icu_properties

Module bidi_data

Data and APIs for supporting specific Bidi properties data in an efficient structure.

Supported properties are:

Bidi_Paired_Bracket

Bidi_Paired_Bracket_Type

Bidi_Mirrored

Bidi_Mirroring_Glyph

Structs

BidiAuxiliaryProperties

A wrapper around certain Bidi properties data. Can be obtained via bidi_auxiliary_properties() and related getters.

BidiAuxiliaryPropertiesBorrowed

A borrowed wrapper around Bidi properties data, returned by BidiAuxiliaryProperties::as_borrowed(). More efficient to query.

BidiMirroringProperties

This struct represents the properties Bidi_Mirrored and Bidi_Mirroring_Glyph. If Bidi_Mirroring_Glyph is not defined for a code point, then the value in the struct is None.

Enums

BidiPairingProperties

The enum represents Bidi_Paired_Bracket_Type, the char represents Bidi_Paired_Bracket. Bidi_Paired_Bracket has a value of None when Bidi_Paired_Bracket_Type is None.

Functions

bidi_auxiliary_properties

Creates a BidiAuxiliaryPropertiesV1 struct that represents the data for certain Bidi properties.

load_bidi_auxiliary_properties_unstable

A version of bidi_auxiliary_properties that uses custom data provided by a DataProvider.

load_bidi_auxiliary_properties_with_any_provider

A version of bidi_auxiliary_properties that uses custom data provided by an AnyProvider.