icu_properties

Enum GeneralCategory

#[repr(u8)]
pub enum GeneralCategory {

    Unassigned = 0,
    UppercaseLetter = 1,
    LowercaseLetter = 2,
    TitlecaseLetter = 3,
    ModifierLetter = 4,
    OtherLetter = 5,
    NonspacingMark = 6,
    SpacingMark = 8,
    EnclosingMark = 7,
    DecimalNumber = 9,
    LetterNumber = 10,
    OtherNumber = 11,
    SpaceSeparator = 12,
    LineSeparator = 13,
    ParagraphSeparator = 14,
    Control = 15,
    Format = 16,
    PrivateUse = 17,
    Surrogate = 18,
    DashPunctuation = 19,
    OpenPunctuation = 20,
    ClosePunctuation = 21,
    ConnectorPunctuation = 22,
    InitialPunctuation = 28,
    FinalPunctuation = 29,
    OtherPunctuation = 23,
    MathSymbol = 24,
    CurrencySymbol = 25,
    ModifierSymbol = 26,
    OtherSymbol = 27,

}

Enumerated property General_Category.

General_Category specifies the most general classification of a code point, usually determined based on the primary characteristic of the assigned character. For example, is the character a letter, a mark, a number, punctuation, or a symbol, and if so, of what type?

GeneralCategory only supports specific subcategories (eg UppercaseLetter). It does not support grouped categories (eg Letter). For grouped categories, use GeneralCategoryGroup.

Variants

Unassigned = 0

(Cn) A reserved unassigned code point or a noncharacter

UppercaseLetter = 1

(Lu) An uppercase letter

LowercaseLetter = 2

(Ll) A lowercase letter

TitlecaseLetter = 3

(Lt) A digraphic letter, with first part uppercase

ModifierLetter = 4

(Lm) A modifier letter

OtherLetter = 5

(Lo) Other letters, including syllables and ideographs

NonspacingMark = 6

(Mn) A nonspacing combining mark (zero advance width)

SpacingMark = 8

(Mc) A spacing combining mark (positive advance width)

EnclosingMark = 7

(Me) An enclosing combining mark

DecimalNumber = 9

(Nd) A decimal digit

LetterNumber = 10

(Nl) A letterlike numeric character

OtherNumber = 11

(No) A numeric character of other type

SpaceSeparator = 12

(Zs) A space character (of various non-zero widths)

LineSeparator = 13

(Zl) U+2028 LINE SEPARATOR only

ParagraphSeparator = 14

(Zp) U+2029 PARAGRAPH SEPARATOR only

Control = 15

(Cc) A C0 or C1 control code

Format = 16

(Cf) A format control character

PrivateUse = 17

(Co) A private-use character

Surrogate = 18

(Cs) A surrogate code point

DashPunctuation = 19

(Pd) A dash or hyphen punctuation mark

OpenPunctuation = 20

(Ps) An opening punctuation mark (of a pair)

ClosePunctuation = 21

(Pe) A closing punctuation mark (of a pair)

ConnectorPunctuation = 22

(Pc) A connecting punctuation mark, like a tie

InitialPunctuation = 28

(Pi) An initial quotation mark

FinalPunctuation = 29

(Pf) A final quotation mark

OtherPunctuation = 23

(Po) A punctuation mark of other type

MathSymbol = 24

(Sm) A symbol of mathematical use

CurrencySymbol = 25

(Sc) A currency sign

ModifierSymbol = 26

(Sk) A non-letterlike modifier symbol

OtherSymbol = 27

(So) A symbol of other type

Implementations

impl GeneralCategory

pub const fn name_to_enum_mapper() -> PropertyValueNameToEnumMapperBorrowed<'static, GeneralCategory>

Return a PropertyValueNameToEnumMapper, capable of looking up values from strings for the General_Category enumerated property.

โจ Enabled with the compiled_data Cargo feature.

๐ Help choosing a constructor

Example

use icu::properties::GeneralCategory;

let lookup = GeneralCategory::name_to_enum_mapper();
// short name for value
assert_eq!(lookup.get_strict("Lu"), Some(GeneralCategory::UppercaseLetter));
assert_eq!(lookup.get_strict("Pd"), Some(GeneralCategory::DashPunctuation));
// long name for value
assert_eq!(lookup.get_strict("Uppercase_Letter"), Some(GeneralCategory::UppercaseLetter));
assert_eq!(lookup.get_strict("Dash_Punctuation"), Some(GeneralCategory::DashPunctuation));
// name has incorrect casing
assert_eq!(lookup.get_strict("dashpunctuation"), None);
// loose matching of name
assert_eq!(lookup.get_loose("dash-punctuation"), Some(GeneralCategory::DashPunctuation));
// fake property
assert_eq!(lookup.get_loose("Animated_Gif"), None);

pub fn get_name_to_enum_mapper( provider: &(impl DataProvider<GeneralCategoryNameToValueV1Marker> + ?Sized), ) -> Result<PropertyValueNameToEnumMapper<GeneralCategory>, PropertiesError>

A version of GeneralCategory::name_to_enum_mapper() that uses custom data provided by a DataProvider.

๐ Help choosing a constructor

pub const fn enum_to_short_name_mapper() -> PropertyEnumToValueNameLinearMapperBorrowed<'static, GeneralCategory>

Return a PropertyEnumToValueNameLinearMapper, capable of looking up short names for values of the General_Category enumerated property.

โจ Enabled with the compiled_data Cargo feature.

๐ Help choosing a constructor

Example

use icu::properties::GeneralCategory;

let lookup = GeneralCategory::enum_to_short_name_mapper();
assert_eq!(lookup.get(GeneralCategory::UppercaseLetter), Some("Lu"));
assert_eq!(lookup.get(GeneralCategory::DashPunctuation), Some("Pd"));
assert_eq!(lookup.get(GeneralCategory::FinalPunctuation), Some("Pf"));

pub fn get_enum_to_short_name_mapper( provider: &(impl DataProvider<GeneralCategoryValueToShortNameV1Marker> + ?Sized), ) -> Result<PropertyEnumToValueNameLinearMapper<GeneralCategory>, PropertiesError>

A version of GeneralCategory::enum_to_short_name_mapper() that uses custom data provided by a DataProvider.

๐ Help choosing a constructor

pub const fn enum_to_long_name_mapper() -> PropertyEnumToValueNameLinearMapperBorrowed<'static, GeneralCategory>

Return a PropertyEnumToValueNameLinearMapper, capable of looking up long names for values of the General_Category enumerated property.

โจ Enabled with the compiled_data Cargo feature.

๐ Help choosing a constructor

Example

use icu::properties::GeneralCategory;

let lookup = GeneralCategory::enum_to_long_name_mapper();
assert_eq!(lookup.get(GeneralCategory::UppercaseLetter), Some("Uppercase_Letter"));
assert_eq!(lookup.get(GeneralCategory::DashPunctuation), Some("Dash_Punctuation"));
assert_eq!(lookup.get(GeneralCategory::FinalPunctuation), Some("Final_Punctuation"));

pub fn get_enum_to_long_name_mapper( provider: &(impl DataProvider<GeneralCategoryValueToLongNameV1Marker> + ?Sized), ) -> Result<PropertyEnumToValueNameLinearMapper<GeneralCategory>, PropertiesError>

A version of GeneralCategory::enum_to_long_name_mapper() that uses custom data provided by a DataProvider.

๐ Help choosing a constructor