Module icu_properties::maps

The functions in this module return a CodePointTrie representing, for each code point in the entire range of code points, the property values for a particular Unicode property.

The descriptions of most properties are taken from TR44, the documentation for the Unicode Character Database.

Functions

get_canonical_combining_class

Return a CodePointTrie for the Canonical_Combining_Class Unicode property. See CanonicalCombiningClass.

get_east_asian_width

Return a CodePointTrie for the East_Asian_Width Unicode enumerated property. See EastAsianWidth.

get_general_category

Return a CodePointTrie for the General_Category Unicode enumerated property. See GeneralCategory.

get_grapheme_cluster_break

Return a CodePointTrie for the Grapheme_Cluster_Break Unicode enumerated property. See GraphemeClusterBreak.

get_line_break

Return a CodePointTrie for the Line_Break Unicode enumerated property. See LineBreak.

get_script

Return a CodePointTrie for the Script Unicode enumerated property. See Script.

get_sentence_break

Return a CodePointTrie for the Sentence_Break Unicode enumerated property. See SentenceBreak.

get_word_break

Return a CodePointTrie for the Word_Break Unicode enumerated property. See WordBreak.

Type Definitions

CodePointMapResult

TODO(#1239): Finalize this API.