icu_properties

Module names

Macros

impl_value_getter 🔒

Structs

PropertyNamesLong

A struct capable of looking up a property name from a value Access its data by calling Self::as_borrowed() and using the methods on PropertyNamesLongBorrowed.

PropertyNamesLongBorrowed

A borrowed wrapper around property value name-to-enum data, returned by PropertyNamesLong::as_borrowed(). More efficient to query.

PropertyNamesShort

A struct capable of looking up a property name from a value Access its data by calling Self::as_borrowed() and using the methods on PropertyNamesShortBorrowed.

PropertyNamesShortBorrowed

A borrowed wrapper around property value name-to-enum data, returned by PropertyNamesShort::as_borrowed(). More efficient to query.

PropertyParser

A struct capable of looking up a property value from a string name. Access its data by calling Self::as_borrowed() and using the methods on PropertyParserBorrowed.

PropertyParserBorrowed

A borrowed wrapper around property value name-to-enum data, returned by PropertyParser::as_borrowed(). More efficient to query.

Traits

NamedEnumeratedProperty

A property whose value names can be represented as strings.

ParseableEnumeratedProperty

A property whose value names can be parsed from strings.

PropertyEnumToValueNameLookup

Functions

get_loose_u16 🔒

Avoid monomorphizing multiple copies of this function

get_strict_u16 🔒

Avoid monomorphizing multiple copies of this function