icu_properties::props

Enum GeneralCategory

#[repr(u8)]
pub enum GeneralCategory {

    Unassigned = 0,
    UppercaseLetter = 1,
    LowercaseLetter = 2,
    TitlecaseLetter = 3,
    ModifierLetter = 4,
    OtherLetter = 5,
    NonspacingMark = 6,
    EnclosingMark = 7,
    SpacingMark = 8,
    DecimalNumber = 9,
    LetterNumber = 10,
    OtherNumber = 11,
    SpaceSeparator = 12,
    LineSeparator = 13,
    ParagraphSeparator = 14,
    Control = 15,
    Format = 16,
    PrivateUse = 17,
    Surrogate = 18,
    DashPunctuation = 19,
    OpenPunctuation = 20,
    ClosePunctuation = 21,
    ConnectorPunctuation = 22,
    OtherPunctuation = 23,
    MathSymbol = 24,
    CurrencySymbol = 25,
    ModifierSymbol = 26,
    OtherSymbol = 27,
    InitialPunctuation = 28,
    FinalPunctuation = 29,

}

Enumerated property General_Category.

General_Category specifies the most general classification of a code point, usually determined based on the primary characteristic of the assigned character. For example, is the character a letter, a mark, a number, punctuation, or a symbol, and if so, of what type?

GeneralCategory only supports specific subcategories (eg UppercaseLetter). It does not support grouped categories (eg Letter). For grouped categories, use GeneralCategoryGroup.

Example

use icu::properties::{CodePointMapData, props::GeneralCategory};

assert_eq!(
    CodePointMapData::<GeneralCategory>::new().get('木'),
    GeneralCategory::OtherLetter
); // U+6728
assert_eq!(
    CodePointMapData::<GeneralCategory>::new().get('🎃'),
    GeneralCategory::OtherSymbol
); // U+1F383 JACK-O-LANTERN

Variants

Unassigned = 0

UppercaseLetter = 1

LowercaseLetter = 2

TitlecaseLetter = 3

ModifierLetter = 4

OtherLetter = 5

NonspacingMark = 6

EnclosingMark = 7

SpacingMark = 8

DecimalNumber = 9

LetterNumber = 10

OtherNumber = 11

SpaceSeparator = 12

LineSeparator = 13

ParagraphSeparator = 14

Control = 15

Format = 16

PrivateUse = 17

Surrogate = 18

DashPunctuation = 19

OpenPunctuation = 20

ClosePunctuation = 21

ConnectorPunctuation = 22

OtherPunctuation = 23

MathSymbol = 24

CurrencySymbol = 25

ModifierSymbol = 26

OtherSymbol = 27

InitialPunctuation = 28

FinalPunctuation = 29

Implementations

impl GeneralCategory

pub const Cn: Self = Self::Unassigned

pub const Lu: Self = Self::UppercaseLetter

pub const Ll: Self = Self::LowercaseLetter

pub const Lt: Self = Self::TitlecaseLetter

pub const Lm: Self = Self::ModifierLetter

pub const Lo: Self = Self::OtherLetter

pub const Mn: Self = Self::NonspacingMark

pub const Me: Self = Self::EnclosingMark

pub const Mc: Self = Self::SpacingMark

pub const Nd: Self = Self::DecimalNumber

pub const Digit: Self = Self::DecimalNumber

pub const Nl: Self = Self::LetterNumber

pub const No: Self = Self::OtherNumber

pub const Zs: Self = Self::SpaceSeparator

pub const Zl: Self = Self::LineSeparator

pub const Zp: Self = Self::ParagraphSeparator

pub const Cc: Self = Self::Control

pub const Cntrl: Self = Self::Control

pub const Cf: Self = Self::Format

pub const Co: Self = Self::PrivateUse

pub const Cs: Self = Self::Surrogate

pub const Pd: Self = Self::DashPunctuation

pub const Ps: Self = Self::OpenPunctuation

pub const Pe: Self = Self::ClosePunctuation

pub const Pc: Self = Self::ConnectorPunctuation

pub const Po: Self = Self::OtherPunctuation

pub const Sm: Self = Self::MathSymbol

pub const Sc: Self = Self::CurrencySymbol

pub const Sk: Self = Self::ModifierSymbol

pub const So: Self = Self::OtherSymbol

pub const Pi: Self = Self::InitialPunctuation

pub const Pf: Self = Self::FinalPunctuation

impl GeneralCategory

pub const ALL_VALUES: &'static [Self]

👎Deprecated

Deprecated