icu_properties::props

Struct BidiClass

pub struct BidiClass(/* private fields */);

Enumerated property Bidi_Class

These are the categories required by the Unicode Bidirectional Algorithm. For the property values, see Bidirectional Class Values. For more information, see Unicode Standard Annex #9.

Example

use icu::properties::{props::BidiClass, CodePointMapData};

assert_eq!(
    CodePointMapData::<BidiClass>::new().get('y'),
    BidiClass::LeftToRight
); // U+0079
assert_eq!(
    CodePointMapData::<BidiClass>::new().get('ع'),
    BidiClass::ArabicLetter
); // U+0639

Implementations

impl BidiClass

pub const fn to_icu4c_value(self) -> u8

Returns an ICU4C UBidiClass value.

pub const fn from_icu4c_value(value: u8) -> Self

Constructor from an ICU4C UBidiClass value.

impl BidiClass

pub const LeftToRight: BidiClass

(L) any strong left-to-right character

pub const RightToLeft: BidiClass

(R) any strong right-to-left (non-Arabic-type) character

pub const EuropeanNumber: BidiClass

(EN) any ASCII digit or Eastern Arabic-Indic digit

pub const EuropeanSeparator: BidiClass

(ES) plus and minus signs

pub const EuropeanTerminator: BidiClass

(ET) a terminator in a numeric format context, includes currency signs

pub const ArabicNumber: BidiClass

(AN) any Arabic-Indic digit

pub const CommonSeparator: BidiClass

(CS) commas, colons, and slashes

pub const ParagraphSeparator: BidiClass

(B) various newline characters

pub const SegmentSeparator: BidiClass

(S) various segment-related control codes

pub const WhiteSpace: BidiClass

(WS) spaces

pub const OtherNeutral: BidiClass

(ON) most other symbols and punctuation marks

pub const LeftToRightEmbedding: BidiClass

(LRE) U+202A: the LR embedding control

pub const LeftToRightOverride: BidiClass

(LRO) U+202D: the LR override control

pub const ArabicLetter: BidiClass

(AL) any strong right-to-left (Arabic-type) character

pub const RightToLeftEmbedding: BidiClass

(RLE) U+202B: the RL embedding control

pub const RightToLeftOverride: BidiClass

(RLO) U+202E: the RL override control

pub const PopDirectionalFormat: BidiClass

(PDF) U+202C: terminates an embedding or override control

pub const NonspacingMark: BidiClass

(NSM) any nonspacing mark

pub const BoundaryNeutral: BidiClass

(BN) most format characters, control codes, or noncharacters

pub const FirstStrongIsolate: BidiClass

(FSI) U+2068: the first strong isolate control

pub const LeftToRightIsolate: BidiClass

(LRI) U+2066: the LR isolate control

pub const RightToLeftIsolate: BidiClass

(RLI) U+2067: the RL isolate control

pub const PopDirectionalIsolate: BidiClass

(PDI) U+2069: terminates an isolate control

pub const ALL_VALUES: &'static [BidiClass]

All possible values of this enum in the Unicode version from this ICU4X release.