icu_properties::props

Struct EmojiComponent

#[non_exhaustive]
pub struct EmojiComponent;

Characters used in emoji sequences that normally do not appear on emoji keyboards as separate choices, such as base characters for emoji keycaps.

Example

use icu::properties::CodePointSetData;
use icu::properties::props::EmojiComponent;

let emoji_component = CodePointSetData::new::<EmojiComponent>();

assert!(emoji_component.contains('๐น'));  // U+1F1F9 REGIONAL INDICATOR SYMBOL LETTER T
assert!(emoji_component.contains('\u{20E3}'));  // COMBINING ENCLOSING KEYCAP
assert!(emoji_component.contains('7'));
assert!(!emoji_component.contains('T'));