icu_properties::props

Struct JoiningGroup

pub struct JoiningGroup(/* private fields */);

Enumerated property Joining_Group.

See Section 9.2, Arabic Joining Groups in The Unicode Standard for the summary of each property value.

use icu::properties::{props::JoiningGroup, CodePointMapData};

assert_eq!(
    CodePointMapData::<JoiningGroup>::new().get('ع'),
    JoiningGroup::Ain,
); // U+0639: Arabic Letter Ain
assert_eq!(
    CodePointMapData::<JoiningGroup>::new().get('ظ'),
    JoiningGroup::Tah,
); // U+0638: Arabic Letter Zah

Implementations

impl JoiningGroup

pub const fn to_icu4c_value(self) -> u8

Returns an ICU4C UJoiningType value.

pub const fn from_icu4c_value(value: u8) -> Self

Constructor from an ICU4C UJoiningType value.

impl JoiningGroup

pub const NoJoiningGroup: JoiningGroup

pub const Ain: JoiningGroup

pub const Alaph: JoiningGroup

pub const Alef: JoiningGroup

pub const Beh: JoiningGroup

pub const Beth: JoiningGroup

pub const Dal: JoiningGroup

pub const DalathRish: JoiningGroup

pub const E: JoiningGroup

pub const Feh: JoiningGroup

pub const FinalSemkath: JoiningGroup

pub const Gaf: JoiningGroup

pub const Gamal: JoiningGroup

pub const Hah: JoiningGroup

pub const TehMarbutaGoal: JoiningGroup

pub const He: JoiningGroup

pub const Heh: JoiningGroup

pub const HehGoal: JoiningGroup

pub const Heth: JoiningGroup

pub const Kaf: JoiningGroup

pub const Kaph: JoiningGroup

pub const KnottedHeh: JoiningGroup

pub const Lam: JoiningGroup

pub const Lamadh: JoiningGroup

pub const Meem: JoiningGroup

pub const Mim: JoiningGroup

pub const Noon: JoiningGroup

pub const Nun: JoiningGroup

pub const Pe: JoiningGroup

pub const Qaf: JoiningGroup

pub const Qaph: JoiningGroup

pub const Reh: JoiningGroup

pub const ReversedPe: JoiningGroup

pub const Sad: JoiningGroup

pub const Sadhe: JoiningGroup

pub const Seen: JoiningGroup

pub const Semkath: JoiningGroup

pub const Shin: JoiningGroup

pub const SwashKaf: JoiningGroup

pub const SyriacWaw: JoiningGroup

pub const Tah: JoiningGroup

pub const Taw: JoiningGroup

pub const TehMarbuta: JoiningGroup

pub const Teth: JoiningGroup

pub const Waw: JoiningGroup

pub const Yeh: JoiningGroup

pub const YehBarree: JoiningGroup

pub const YehWithTail: JoiningGroup

pub const Yudh: JoiningGroup

pub const YudhHe: JoiningGroup

pub const Zain: JoiningGroup

pub const Fe: JoiningGroup

pub const Khaph: JoiningGroup

pub const Zhain: JoiningGroup

pub const BurushaskiYehBarree: JoiningGroup

pub const FarsiYeh: JoiningGroup

pub const Nya: JoiningGroup

pub const RohingyaYeh: JoiningGroup

pub const ManichaeanAleph: JoiningGroup

pub const ManichaeanAyin: JoiningGroup

pub const ManichaeanBeth: JoiningGroup

pub const ManichaeanDaleth: JoiningGroup

pub const ManichaeanDhamedh: JoiningGroup

pub const ManichaeanFive: JoiningGroup

pub const ManichaeanGimel: JoiningGroup

pub const ManichaeanHeth: JoiningGroup

pub const ManichaeanHundred: JoiningGroup

pub const ManichaeanKaph: JoiningGroup

pub const ManichaeanLamedh: JoiningGroup

pub const ManichaeanMem: JoiningGroup

pub const ManichaeanNun: JoiningGroup

pub const ManichaeanOne: JoiningGroup

pub const ManichaeanPe: JoiningGroup

pub const ManichaeanQoph: JoiningGroup

pub const ManichaeanResh: JoiningGroup

pub const ManichaeanSadhe: JoiningGroup

pub const ManichaeanSamekh: JoiningGroup

pub const ManichaeanTaw: JoiningGroup

pub const ManichaeanTen: JoiningGroup

pub const ManichaeanTeth: JoiningGroup

pub const ManichaeanThamedh: JoiningGroup

pub const ManichaeanTwenty: JoiningGroup

pub const ManichaeanWaw: JoiningGroup

pub const ManichaeanYodh: JoiningGroup

pub const ManichaeanZayin: JoiningGroup

pub const StraightWaw: JoiningGroup

pub const AfricanFeh: JoiningGroup

pub const AfricanNoon: JoiningGroup

pub const AfricanQaf: JoiningGroup

pub const MalayalamBha: JoiningGroup

pub const MalayalamJa: JoiningGroup

pub const MalayalamLla: JoiningGroup

pub const MalayalamLlla: JoiningGroup

pub const MalayalamNga: JoiningGroup

pub const MalayalamNna: JoiningGroup

pub const MalayalamNnna: JoiningGroup

pub const MalayalamNya: JoiningGroup

pub const MalayalamRa: JoiningGroup

pub const MalayalamSsa: JoiningGroup

pub const MalayalamTta: JoiningGroup

pub const HanifiRohingyaKinnaYa: JoiningGroup

pub const HanifiRohingyaPa: JoiningGroup

pub const ThinYeh: JoiningGroup

pub const VerticalTail: JoiningGroup

pub const KashmiriYeh: JoiningGroup

pub const ThinNoon: JoiningGroup

pub const ALL_VALUES: &'static [JoiningGroup]

All possible values of this enum in the Unicode version from this ICU4X release.