icu_properties::props

Trait EmojiSet

pub trait EmojiSet: Sealed {
    const NAME: &'static [u8];
    const SHORT_NAME: &'static [u8];
}

An Emoji set as defined by Unicode Technical Standard #51.

๐ซ This trait is sealed; it cannot be implemented by user code. If an API requests an item that implements this trait, please consider using a type from the implementors listed below.

Required Associated Constants

const NAME: &'static [u8]

The name of this property

const SHORT_NAME: &'static [u8]

The abbreviated name of this property, if it exists, otherwise the name

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.