icu_properties::props

Trait EnumeratedProperty

pub trait EnumeratedProperty: Sealed + TrieValue {
    const NAME: &'static [u8];
    const SHORT_NAME: &'static [u8];

    // Provided method
    fn for_char(ch: char) -> Self { ... }
}

A Unicode character property that assigns a value to each code point.

The descriptions of most properties are taken from TR44, the documentation for the Unicode Character Database.

๐ซ This trait is sealed; it cannot be implemented by user code. If an API requests an item that implements this trait, please consider using a type from the implementors listed below.

Required Associated Constants

const NAME: &'static [u8]

The name of this property

const SHORT_NAME: &'static [u8]

The abbreviated name of this property, if it exists, otherwise the name

Provided Methods

fn for_char(ch: char) -> Self

Convenience method for CodePointMapData::new().get(ch)

โจ Enabled with the compiled_data Cargo feature.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.