icu_properties

Module provider

🚧 [Unstable] Data provider struct definitions for this ICU4X component.

🚧 This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Read more about data providers: icu_provider

Re-exports

pub use self::names::GeneralCategoryMaskNameToValueV1Marker;

Modules

bidi_data

🚧 [Unstable] Data provider struct definitions for this ICU4X component.

names

🚧 [Unstable] Property names-related data for this component

Structs

AlnumV1Marker

Data marker for the ‘alnum’ Unicode property

AlphabeticV1Marker

Data marker for the ‘Alpha’ Unicode property

AsciiHexDigitV1Marker

Data marker for the ‘AHex’ Unicode property

Baked

Baked data

BasicEmojiV1Marker

Data marker for the ‘Basic_Emoji’ Unicode property

BidiClassNameToValueV1Marker

Data marker for parsing the names of the values of the ‘bc’ Unicode property

BidiClassV1Marker

Data marker for the ‘bc’ Unicode property

BidiClassValueToLongNameV1Marker

Data marker for producing long names of the values of the ‘bc’ Unicode property

BidiClassValueToShortNameV1Marker

Data marker for producing short names of the values of the ‘bc’ Unicode property

BidiControlV1Marker

Data marker for the ‘Bidi_C’ Unicode property

BidiMirroredV1Marker

Data marker for the ‘Bidi_M’ Unicode property

BlankV1Marker

Data marker for the ‘blank’ Unicode property

CanonicalCombiningClassNameToValueV1Marker

Data marker for parsing the names of the values of the ‘ccc’ Unicode property

CanonicalCombiningClassV1Marker

Data marker for the ‘ccc’ Unicode property

CanonicalCombiningClassValueToLongNameV1Marker

Data marker for producing long names of the values of the ‘ccc’ Unicode property

CanonicalCombiningClassValueToShortNameV1Marker

Data marker for producing short names of the values of the ‘ccc’ Unicode property

CaseIgnorableV1Marker

Data marker for the ‘CI’ Unicode property

CaseSensitiveV1Marker

Data marker for the ‘Sensitive’ Unicode property

CasedV1Marker

Data marker for the ‘Cased’ Unicode property

ChangesWhenCasefoldedV1Marker

Data marker for the ‘CWCF’ Unicode property

ChangesWhenCasemappedV1Marker

Data marker for the ‘CWCM’ Unicode property

ChangesWhenLowercasedV1Marker

Data marker for the ‘CWL’ Unicode property

ChangesWhenNfkcCasefoldedV1Marker

Data marker for the ‘CWKCF’ Unicode property

ChangesWhenTitlecasedV1Marker

Data marker for the ‘CWT’ Unicode property

ChangesWhenUppercasedV1Marker

Data marker for the ‘CWU’ Unicode property

DashV1Marker

Data marker for the ‘Dash’ Unicode property

DefaultIgnorableCodePointV1Marker

Data marker for the ‘DI’ Unicode property

DeprecatedV1Marker

Data marker for the ‘Dep’ Unicode property

DiacriticV1Marker

Data marker for the ‘Dia’ Unicode property

EastAsianWidthNameToValueV1Marker

Data marker for parsing the names of the values of the ‘ea’ Unicode property

EastAsianWidthV1Marker

Data marker for the ‘ea’ Unicode property

EastAsianWidthValueToLongNameV1Marker

Data marker for producing long names of the values of the ‘ea’ Unicode property

EastAsianWidthValueToShortNameV1Marker

Data marker for producing short names of the values of the ‘ea’ Unicode property

EmojiComponentV1Marker

Data marker for the ‘EComp’ Unicode property

EmojiModifierBaseV1Marker

Data marker for the ‘EBase’ Unicode property

EmojiModifierV1Marker

Data marker for the ‘EMod’ Unicode property

EmojiPresentationV1Marker

Data marker for the ‘EPres’ Unicode property

EmojiV1Marker

Data marker for the ‘Emoji’ Unicode property

ExemplarCharactersAuxiliaryV1Marker

Data marker for the ‘exemplarchars/auxiliary’ Unicode property

ExemplarCharactersIndexV1Marker

Data marker for the ‘exemplarchars/index’ Unicode property

ExemplarCharactersMainV1Marker

Data marker for the ‘exemplarchars/main’ Unicode property

ExemplarCharactersNumbersV1Marker

Data marker for the ‘exemplarchars/numbers’ Unicode property

ExemplarCharactersPunctuationV1Marker

Data marker for the ‘exemplarchars/punctuation’ Unicode property

ExtendedPictographicV1Marker

Data marker for the ‘ExtPict’ Unicode property

ExtenderV1Marker

Data marker for the ‘Ext’ Unicode property

FullCompositionExclusionV1Marker

Data marker for the ‘Comp_Ex’ Unicode property

GeneralCategoryNameToValueV1Marker

Data marker for parsing the names of the values of the ‘gc’ Unicode property

GeneralCategoryV1Marker

Data marker for the ‘gc’ Unicode property

GeneralCategoryValueToLongNameV1Marker

Data marker for producing long names of the values of the ‘gc’ Unicode property

GeneralCategoryValueToShortNameV1Marker

Data marker for producing short names of the values of the ‘gc’ Unicode property

GraphV1Marker

Data marker for the ‘graph’ Unicode property

GraphemeBaseV1Marker

Data marker for the ‘Gr_Base’ Unicode property

GraphemeClusterBreakNameToValueV1Marker

Data marker for parsing the names of the values of the ‘GCB’ Unicode property

GraphemeClusterBreakV1Marker

Data marker for the ‘GCB’ Unicode property

GraphemeClusterBreakValueToLongNameV1Marker

Data marker for producing long names of the values of the ‘GCB’ Unicode property

GraphemeClusterBreakValueToShortNameV1Marker

Data marker for producing short names of the values of the ‘GCB’ Unicode property

GraphemeExtendV1Marker

Data marker for the ‘Gr_Ext’ Unicode property

GraphemeLinkV1Marker

Data marker for the ‘Gr_Link’ Unicode property

HangulSyllableTypeNameToValueV1Marker

Data marker for parsing the names of the values of the ‘hst’ Unicode property

HangulSyllableTypeV1Marker

Data marker for the ‘hst’ Unicode property

HangulSyllableTypeValueToLongNameV1Marker

Data marker for producing long names of the values of the ‘hst’ Unicode property

HangulSyllableTypeValueToShortNameV1Marker

Data marker for producing short names of the values of the ‘hst’ Unicode property

HexDigitV1Marker

Data marker for the ‘Hex’ Unicode property

HyphenV1Marker

Data marker for the ‘Hyphen’ Unicode property

IdContinueV1Marker

Data marker for the ‘IDC’ Unicode property

IdStartV1Marker

Data marker for the ‘IDS’ Unicode property

IdeographicV1Marker

Data marker for the ‘Ideo’ Unicode property

IdsBinaryOperatorV1Marker

Data marker for the ‘IDSB’ Unicode property

IdsTrinaryOperatorV1Marker

Data marker for the ‘IDST’ Unicode property

IndicSyllabicCategoryNameToValueV1Marker

Data marker for parsing the names of the values of the ‘InSC’ Unicode property

IndicSyllabicCategoryV1Marker

Data marker for the ‘InSC’ Unicode property

IndicSyllabicCategoryValueToLongNameV1Marker

Data marker for producing long names of the values of the ‘InSC’ Unicode property

IndicSyllabicCategoryValueToShortNameV1Marker

Data marker for producing short names of the values of the ‘InSC’ Unicode property

JoinControlV1Marker

Data marker for the ‘Join_C’ Unicode property

JoiningTypeNameToValueV1Marker

Data marker for parsing the names of the values of the ‘jt’ Unicode property

JoiningTypeV1Marker

Data marker for the ‘jt’ Unicode property

JoiningTypeValueToLongNameV1Marker

Data marker for producing long names of the values of the ‘jt’ Unicode property

JoiningTypeValueToShortNameV1Marker

Data marker for producing short names of the values of the ‘jt’ Unicode property

LineBreakNameToValueV1Marker

Data marker for parsing the names of the values of the ‘lb’ Unicode property

LineBreakV1Marker

Data marker for the ‘lb’ Unicode property

LineBreakValueToLongNameV1Marker

Data marker for producing long names of the values of the ‘lb’ Unicode property

LineBreakValueToShortNameV1Marker

Data marker for producing short names of the values of the ‘lb’ Unicode property

LogicalOrderExceptionV1Marker

Data marker for the ‘LOE’ Unicode property

LowercaseV1Marker

Data marker for the ‘Lower’ Unicode property

MathV1Marker

Data marker for the ‘Math’ Unicode property

NfcInertV1Marker

Data marker for the ‘nfcinert’ Unicode property

NfdInertV1Marker

Data marker for the ‘nfdinert’ Unicode property

NfkcInertV1Marker

Data marker for the ‘nfkcinert’ Unicode property

NfkdInertV1Marker

Data marker for the ‘nfkdinert’ Unicode property

NoncharacterCodePointV1Marker

Data marker for the ‘NChar’ Unicode property

PatternSyntaxV1Marker

Data marker for the ‘Pat_Syn’ Unicode property

PatternWhiteSpaceV1Marker

Data marker for the ‘Pat_WS’ Unicode property

PrependedConcatenationMarkV1Marker

Data marker for the ‘PCM’ Unicode property

PrintV1Marker

Data marker for the ‘print’ Unicode property

QuotationMarkV1Marker

Data marker for the ‘QMark’ Unicode property

RadicalV1Marker

Data marker for the ‘Radical’ Unicode property

RegionalIndicatorV1Marker

Data marker for the ‘RI’ Unicode property

ScriptNameToValueV1Marker

Data marker for parsing the names of the values of the ‘sc’ Unicode property

ScriptV1Marker

Data marker for the ‘sc’ Unicode property

ScriptValueToLongNameV1Marker

Data marker for producing long names of the values of the ‘sc’ Unicode property

ScriptValueToShortNameV1Marker

Data marker for producing short names of the values of the ‘sc’ Unicode property

ScriptWithExtensionsPropertyV1

A struct that efficiently stores Script and Script_Extensions property data.

ScriptWithExtensionsPropertyV1Marker

Marker type for ScriptWithExtensionsPropertyV1: “props/scx@1”

SegmentStarterV1Marker

Data marker for the ‘segstart’ Unicode property

SentenceBreakNameToValueV1Marker

Data marker for parsing the names of the values of the ‘SB’ Unicode property

SentenceBreakV1Marker

Data marker for the ‘SB’ Unicode property

SentenceBreakValueToLongNameV1Marker

Data marker for producing long names of the values of the ‘SB’ Unicode property

SentenceBreakValueToShortNameV1Marker

Data marker for producing short names of the values of the ‘SB’ Unicode property

SentenceTerminalV1Marker

Data marker for the ‘STerm’ Unicode property

SoftDottedV1Marker

Data marker for the ‘SD’ Unicode property

TerminalPunctuationV1Marker

Data marker for the ‘Term’ Unicode property

UnifiedIdeographV1Marker

Data marker for the ‘UIdeo’ Unicode property

UppercaseV1Marker

Data marker for the ‘Upper’ Unicode property

VariationSelectorV1Marker

Data marker for the ‘VS’ Unicode property

WhiteSpaceV1Marker

Data marker for the ‘WSpace’ Unicode property

WordBreakNameToValueV1Marker

Data marker for parsing the names of the values of the ‘WB’ Unicode property

WordBreakV1Marker

Data marker for the ‘WB’ Unicode property

WordBreakValueToLongNameV1Marker

Data marker for producing long names of the values of the ‘WB’ Unicode property

WordBreakValueToShortNameV1Marker

Data marker for producing short names of the values of the ‘WB’ Unicode property

XdigitV1Marker

Data marker for the ‘xdigit’ Unicode property

XidContinueV1Marker

Data marker for the ‘XIDC’ Unicode property

XidStartV1Marker

Data marker for the ‘XIDS’ Unicode property

Enums

PropertyCodePointMapV1

A map efficiently storing data about individual characters.

PropertyCodePointSetV1

A set of characters which share a particular property value.

PropertyUnicodeSetV1

A set of characters and strings which share a particular property value.

Constants

KEYS

All data keys in this module.