Module icu_properties::provider::bidi_data

🚧 [Unstable] Data provider struct definitions for this ICU4X component.

🚧 This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Read more about data providers: icu_provider

This module provides an efficient storage of data serving the following properties:

Bidi_Paired_Bracket

Bidi_Paired_Bracket_Type

Bidi_Mirrored

Bidi_Mirroring_Glyph

Structs

BidiAuxiliaryPropertiesV1

A data provider struct for properties related to Bidi algorithms, including mirroring and bracket pairing.

BidiAuxiliaryPropertiesV1Marker

Marker type for BidiAuxiliaryPropertiesV1: “props/bidiauxiliaryprops@1”

CheckedBidiPairedBracketTypeULE

ULE type for CheckedBidiPairedBracketType

MirroredPairedBracketDataTryFromError

A u32 serialized value of MirroredPairedBracketData did not encode either a valid Bidi_Mirroring_Glyph or a valid Bidi_Paired_Bracket_Type

Enums

CheckedBidiPairedBracketType

A closed Rust enum representing a closed set of the incoming Bidi_Paired_Bracket_Type property values necessary in the internal representation of MirroredPairedBracketData to satisfy the ULE invariants on valid values.