icu_properties::provider

Enum PropertyUnicodeSetV1

#[non_exhaustive]
pub enum PropertyUnicodeSetV1<'data> {
    CPInversionListStrList(CodePointInversionListAndStringList<'data>),
}

A set of characters and strings which share a particular property value.

๐ง This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

CPInversionListStrList(CodePointInversionListAndStringList<'data>)

A set representing characters in an inversion list, and the strings in a list.