Module icu_properties::provider::key

Resource keys for icu_uniset

Constants

ALL_MAP_KEYS

The set of all resource keys supported by icu_uniset.

ALL_SCRIPT_EXTENSIONS_KEYS

The set of all resource keys supported by icu_uniset.

ALL_SET_KEYS

The set of all resource keys supported by icu_uniset.

ALNUM_V1

ALPHABETIC_V1

ASCII_HEX_DIGIT_V1

BIDI_CLASS

BIDI_CONTROL_V1

BIDI_MIRRORED_V1

BLANK_V1

CANONICAL_COMBINING_CLASS_V1

CASED_V1

CASE_IGNORABLE_V1

CASE_SENSITIVE_V1

CHANGES_WHEN_CASEFOLDED_V1

CHANGES_WHEN_CASEMAPPED_V1

CHANGES_WHEN_LOWERCASED_V1

CHANGES_WHEN_NFKC_CASEFOLDED_V1

CHANGES_WHEN_TITLECASED_V1

CHANGES_WHEN_UPPERCASED_V1

DASH_V1

DEFAULT_IGNORABLE_CODE_POINT_V1

DEPRECATED_V1

DIACRITIC_V1

EAST_ASIAN_WIDTH_V1

EMOJI_COMPONENT_V1

EMOJI_MODIFIER_BASE_V1

EMOJI_MODIFIER_V1

EMOJI_PRESENTATION_V1

EMOJI_V1

EXTENDED_PICTOGRAPHIC_V1

EXTENDER_V1

FULL_COMPOSITION_EXCLUSION_V1

GENERAL_CATEGORY_V1

GRAPHEME_BASE_V1

GRAPHEME_CLUSTER_BREAK_V1

GRAPHEME_EXTEND_V1

GRAPHEME_LINK_V1

GRAPH_V1

HEX_DIGIT_V1

HYPHEN_V1

IDEOGRAPHIC_V1

IDS_BINARY_OPERATOR_V1

IDS_TRINARY_OPERATOR_V1

ID_CONTINUE_V1

ID_START_V1

JOIN_CONTROL_V1

LINE_BREAK_V1

LOGICAL_ORDER_EXCEPTION_V1

LOWERCASE_V1

MATH_V1

NFC_INERT_V1

NFD_INERT_V1

NFKC_INERT_V1

NFKD_INERT_V1

NONCHARACTER_CODE_POINT_V1

PATTERN_SYNTAX_V1

PATTERN_WHITE_SPACE_V1

PREPENDED_CONCATENATION_MARK_V1

PRINT_V1

QUOTATION_MARK_V1

RADICAL_V1

REGIONAL_INDICATOR_V1

SCRIPT_EXTENSIONS_V1

SCRIPT_V1

SEGMENT_STARTER_V1

SENTENCE_BREAK_V1

SENTENCE_TERMINAL_V1

SOFT_DOTTED_V1

TERMINAL_PUNCTUATION_V1

UNIFIED_IDEOGRAPH_V1

UPPERCASE_V1

VARIATION_SELECTOR_V1

WHITE_SPACE_V1

WORD_BREAK_V1

XDIGIT_V1

XID_CONTINUE_V1

XID_START_V1