icu_properties::provider
Module names

🚧 [Unstable] Property names-related data for this component

🚧 This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Read more about data providers: icu_provider

Structs

PropertyEnumToValueNameLinearMap

A mapping of property values to their names. A single instance of this map will only cover either long or short names, determined whilst loading data.

PropertyEnumToValueNameSparseMap

A mapping of property values to their names. A single instance of this map will only cover either long or short names, determined whilst loading data.

PropertyNameLongBidiClassV1

PropertyNameLongBidiClassV1

PropertyNameLongCanonicalCombiningClassV1

PropertyNameLongCanonicalCombiningClassV1

PropertyNameLongEastAsianWidthV1

PropertyNameLongEastAsianWidthV1

PropertyNameLongGeneralCategoryV1

PropertyNameLongGeneralCategoryV1

PropertyNameLongGraphemeClusterBreakV1

PropertyNameLongGraphemeClusterBreakV1

PropertyNameLongHangulSyllableTypeV1

PropertyNameLongHangulSyllableTypeV1

PropertyNameLongIndicSyllabicCategoryV1

PropertyNameLongIndicSyllabicCategoryV1

PropertyNameLongJoiningTypeV1

PropertyNameLongJoiningTypeV1

PropertyNameLongLineBreakV1

PropertyNameLongLineBreakV1

PropertyNameLongScriptV1

PropertyNameLongScriptV1

PropertyNameLongSentenceBreakV1

PropertyNameLongSentenceBreakV1

PropertyNameLongVerticalOrientationV1

PropertyNameLongVerticalOrientationV1

PropertyNameLongWordBreakV1

PropertyNameLongWordBreakV1

PropertyNameParseBidiClassV1

PropertyNameParseBidiClassV1

PropertyNameParseCanonicalCombiningClassV1

PropertyNameParseCanonicalCombiningClassV1

PropertyNameParseEastAsianWidthV1

PropertyNameParseEastAsianWidthV1

PropertyNameParseGeneralCategoryMaskV1

PropertyNameParseGeneralCategoryMaskV1

PropertyNameParseGeneralCategoryV1

PropertyNameParseGeneralCategoryV1

PropertyNameParseGraphemeClusterBreakV1

PropertyNameParseGraphemeClusterBreakV1

PropertyNameParseHangulSyllableTypeV1

PropertyNameParseHangulSyllableTypeV1

PropertyNameParseIndicSyllabicCategoryV1

PropertyNameParseIndicSyllabicCategoryV1

PropertyNameParseJoiningTypeV1

PropertyNameParseJoiningTypeV1

PropertyNameParseLineBreakV1

PropertyNameParseLineBreakV1

PropertyNameParseScriptV1

PropertyNameParseScriptV1

PropertyNameParseSentenceBreakV1

PropertyNameParseSentenceBreakV1

PropertyNameParseVerticalOrientationV1

PropertyNameParseVerticalOrientationV1

PropertyNameParseWordBreakV1

PropertyNameParseWordBreakV1

PropertyNameShortBidiClassV1

PropertyNameShortBidiClassV1

PropertyNameShortCanonicalCombiningClassV1

PropertyNameShortCanonicalCombiningClassV1

PropertyNameShortEastAsianWidthV1

PropertyNameShortEastAsianWidthV1

PropertyNameShortGeneralCategoryV1

PropertyNameShortGeneralCategoryV1

PropertyNameShortGraphemeClusterBreakV1

PropertyNameShortGraphemeClusterBreakV1

PropertyNameShortHangulSyllableTypeV1

PropertyNameShortHangulSyllableTypeV1

PropertyNameShortIndicSyllabicCategoryV1

PropertyNameShortIndicSyllabicCategoryV1

PropertyNameShortJoiningTypeV1

PropertyNameShortJoiningTypeV1

PropertyNameShortLineBreakV1

PropertyNameShortLineBreakV1

PropertyNameShortScriptV1

PropertyNameShortScriptV1

PropertyNameShortSentenceBreakV1

PropertyNameShortSentenceBreakV1

PropertyNameShortVerticalOrientationV1

PropertyNameShortVerticalOrientationV1

PropertyNameShortWordBreakV1

PropertyNameShortWordBreakV1

PropertyScriptToIcuScriptMap

A mapping of property values to their names. A single instance of this map will only cover either long or short names, determined whilst loading data.

PropertyValueNameToEnumMap

A set of characters and strings which share a particular property value.