icu_properties::provider::names

Struct NormalizedPropertyNameStr

pub struct NormalizedPropertyNameStr(/* private fields */);

This is a property name that can be โloose matchedโ as according to PropertyValueAliases.txt

(matched case-insensitively in ASCII, ignoring underscores, whitespace, and hyphens)

This is expected to be ASCII, but we do not rely on this invariant anywhere except during datagen.

The Ord impl will sort things using strict equality, but in such a way that all loose-equal items will sort into the same area, such that a map can be searched for both strict and loose equality.

๐ง This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Examples

Using a NormalizedPropertyNameStr as the key of a ZeroMap:

use icu::properties::provider::names::NormalizedPropertyNameStr;
use zerovec::ZeroMap;

let map: ZeroMap<NormalizedPropertyNameStr, usize> = [
    (NormalizedPropertyNameStr::from_str("A_BC"), 11),
    (NormalizedPropertyNameStr::from_str("dEf"), 22),
    (NormalizedPropertyNameStr::from_str("G_H-I"), 33),
]
.into_iter()
.collect();

let key_approx = NormalizedPropertyNameStr::from_str("AB-C");
let key_exact = NormalizedPropertyNameStr::from_str("A_BC");

// Strict lookup:
assert_eq!(None, map.get_copied(key_approx));
assert_eq!(Some(11), map.get_copied(key_exact));

// Loose lookup:
assert_eq!(Some(11), map.get_copied_by(|u| u.cmp_loose(key_approx)));
assert_eq!(Some(11), map.get_copied_by(|u| u.cmp_loose(key_exact)));

Implementations

impl NormalizedPropertyNameStr

pub fn cmp_loose(&self, other: &Self) -> Ordering

Perform the loose comparison as defined in NormalizedPropertyNameStr.

pub const fn from_str(s: &str) -> &Self

Convert a string reference to a NormalizedPropertyNameStr.

pub const fn cast_ref(value: &UnvalidatedStr) -> &Self

Convert a UnvalidatedStr reference to a NormalizedPropertyNameStr reference.

pub const fn cast_box(value: Box<UnvalidatedStr>) -> Box<Self>

Convert a UnvalidatedStr box to a NormalizedPropertyNameStr box.

pub fn boxed_from_bytes(b: &[u8]) -> Box<Self>

Get a NormalizedPropertyNameStr box from a byte slice.