icu_properties

Module runtime

🚧 [Experimental] This module is experimental and currently crate-private. Let us know if you have a use case for this!

This module contains utilities for working with properties where the specific property in use is not known at compile time.

For regex engines, [crate::sets::load_for_ecma262_unstable()] is a convenient API for working with properties at runtime tailored for the use case of ECMA262-compatible regex engines.

Enums

BinaryProperty 🔒

This type can represent any binary Unicode property.

EnumeratedProperty 🔒

This type can represent any enumerated Unicode property.

MaskProperty 🔒

This type can represent any Unicode mask property.

MiscProperty 🔒

NumericProperty 🔒

This type can represent any numeric Unicode property.

StringBinaryProperty 🔒

This type can represent any binary property over strings.

StringProperty 🔒

This type can represent any Unicode string property.