icu_properties::runtime

Enum BinaryProperty

#[non_exhaustive]
enum BinaryProperty {

    Alnum = 44,
    Alphabetic = 0,
    AsciiHexDigit = 1,
    BidiControl = 2,
    BidiMirrored = 3,
    Blank = 45,
    Cased = 49,
    CaseIgnorable = 50,
    CaseSensitive = 34,
    ChangesWhenCasefolded = 54,
    ChangesWhenCasemapped = 55,
    ChangesWhenLowercased = 51,
    ChangesWhenNfkcCasefolded = 56,
    ChangesWhenTitlecased = 53,
    ChangesWhenUppercased = 52,
    Dash = 4,
    DefaultIgnorableCodePoint = 5,
    Deprecated = 6,
    Diacritic = 7,
    Emoji = 57,
    EmojiComponent = 61,
    EmojiModifier = 59,
    EmojiModifierBase = 60,
    EmojiPresentation = 58,
    ExtendedPictographic = 64,
    Extender = 8,
    FullCompositionExclusion = 9,
    Graph = 46,
    GraphemeBase = 10,
    GraphemeExtend = 11,
    GraphemeLink = 12,
    HexDigit = 13,
    Hyphen = 14,
    IdCompatMathContinue = 65,
    IdCompatMathStart = 66,
    IdContinue = 15,
    Ideographic = 17,
    IdsBinaryOperator = 18,
    IdStart = 16,
    IdsTrinaryOperator = 19,
    IdsUnaryOperator = 67,
    JoinControl = 20,
    LogicalOrderException = 21,
    Lowercase = 22,
    Math = 23,
    ModifierCombiningMark = 68,
    NfcInert = 39,
    NfdInert = 37,
    NfkcInert = 40,
    NfkdInert = 38,
    NoncharacterCodePoint = 24,
    PatternSyntax = 42,
    PatternWhiteSpace = 43,
    PrependedConcatenationMark = 63,
    Print = 47,
    QuotationMark = 25,
    Radical = 26,
    RegionalIndicator = 62,
    SegmentStarter = 41,
    SentenceTerminal = 35,
    SoftDotted = 27,
    TerminalPunctuation = 28,
    UnifiedIdeograph = 29,
    Uppercase = 30,
    VariationSelector = 36,
    WhiteSpace = 31,
    Xdigit = 48,
    XidContinue = 32,
    XidStart = 33,

}

This type can represent any binary Unicode property.

This is intended to be used in situations where the exact unicode property needed is only known at runtime, for example in regex engines.

The values are intended to be identical to ICU4C’s UProperty enum

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Alnum = 44

Alphabetic = 0

AsciiHexDigit = 1

BidiControl = 2

BidiMirrored = 3

Blank = 45

Cased = 49

CaseIgnorable = 50

CaseSensitive = 34

ChangesWhenCasefolded = 54

ChangesWhenCasemapped = 55

ChangesWhenLowercased = 51

ChangesWhenNfkcCasefolded = 56

ChangesWhenTitlecased = 53

ChangesWhenUppercased = 52

Dash = 4

DefaultIgnorableCodePoint = 5

Deprecated = 6

Diacritic = 7

Emoji = 57

EmojiComponent = 61

EmojiModifier = 59

EmojiModifierBase = 60

EmojiPresentation = 58

ExtendedPictographic = 64

Extender = 8

FullCompositionExclusion = 9

Graph = 46

GraphemeBase = 10

GraphemeExtend = 11

GraphemeLink = 12

HexDigit = 13

Hyphen = 14

IdCompatMathContinue = 65

IdCompatMathStart = 66

IdContinue = 15

Ideographic = 17

IdsBinaryOperator = 18

IdStart = 16

IdsTrinaryOperator = 19

IdsUnaryOperator = 67

JoinControl = 20

LogicalOrderException = 21

Lowercase = 22

Math = 23

ModifierCombiningMark = 68

NfcInert = 39

NfdInert = 37

NfkcInert = 40

NfkdInert = 38

NoncharacterCodePoint = 24

PatternSyntax = 42

PatternWhiteSpace = 43

PrependedConcatenationMark = 63

Print = 47

QuotationMark = 25

Radical = 26

RegionalIndicator = 62

SegmentStarter = 41

SentenceTerminal = 35

SoftDotted = 27

TerminalPunctuation = 28

UnifiedIdeograph = 29

Uppercase = 30

VariationSelector = 36

WhiteSpace = 31

Xdigit = 48

XidContinue = 32

XidStart = 33