icu_properties::runtime

Enum EnumeratedProperty

#[non_exhaustive]
enum EnumeratedProperty {

    BidiClass = 4_096,
    BidiPairedBracketType = 4_117,
    Block = 4_097,
    CombiningClass = 4_098,
    DecompositionType = 4_099,
    EastAsianWidth = 4_100,
    GeneralCategory = 4_101,
    GraphemeClusterBreak = 4_114,
    HangulSyllableType = 4_107,
    IndicConjunctBreak = 4_122,
    IndicPositionalCategory = 4_118,
    IndicSyllabicCategory = 4_119,
    JoiningGroup = 4_102,
    JoiningType = 4_103,
    LeadCanonicalCombiningClass = 4_112,
    LineBreak = 4_104,
    NFCQuickCheck = 4_110,
    NFDQuickCheck = 4_108,
    NFKCQuickCheck = 4_111,
    NFKDQuickCheck = 4_109,
    NumericType = 4_105,
    Script = 4_106,
    SentenceBreak = 4_115,
    TrailCanonicalCombiningClass = 4_113,
    VerticalOrientation = 4_120,
    WordBreak = 4_116,

}

This type can represent any enumerated Unicode property.

This is intended to be used in situations where the exact unicode property needed is only known at runtime, for example in regex engines.

The values are intended to be identical to ICU4C’s UProperty enum

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

BidiClass = 4_096

BidiPairedBracketType = 4_117

Block = 4_097

CombiningClass = 4_098

DecompositionType = 4_099

EastAsianWidth = 4_100

GeneralCategory = 4_101

GraphemeClusterBreak = 4_114

HangulSyllableType = 4_107

IndicConjunctBreak = 4_122

IndicPositionalCategory = 4_118

IndicSyllabicCategory = 4_119

JoiningGroup = 4_102

JoiningType = 4_103

LeadCanonicalCombiningClass = 4_112

LineBreak = 4_104

NFCQuickCheck = 4_110

NFDQuickCheck = 4_108

NFKCQuickCheck = 4_111

NFKDQuickCheck = 4_109

NumericType = 4_105

Script = 4_106

SentenceBreak = 4_115

TrailCanonicalCombiningClass = 4_113

VerticalOrientation = 4_120

WordBreak = 4_116