icu_properties::runtime

Struct UnicodeProperty

pub struct UnicodeProperty(pub u32);

This type can represent any Unicode property.

This is intended to be used in situations where the exact unicode property needed is only known at runtime, for example in regex engines.

The values are intended to be identical to ICU4C’s UProperty enum

Tuple Fields
0: u32
Implementations

impl UnicodeProperty

pub const Alphabetic: Self

Binary property Alphabetic

pub const AsciiHexDigit: Self

Binary property ASCII_Hex_Digit

pub const BidiControl: Self

Binary property Bidi_Control

pub const BidiMirrored: Self

Binary property Bidi_Mirrored

pub const Dash: Self

Binary property Dash

pub const DefaultIgnorableCodePoint: Self

Binary property Default_Ignorable_Code_Point

pub const Deprecated: Self

Binary property Deprecated

pub const Diacritic: Self

Binary property Diacritic

pub const Extender: Self

Binary property Extender

pub const FullCompositionExclusion: Self

Binary property Full_Composition_Exclusion

pub const GraphemeBase: Self

Binary property Grapheme_Base

pub const GraphemeExtend: Self

Binary property Grapheme_Extend

pub const GraphemeLink: Self

Binary property Grapheme_Link

pub const HexDigit: Self

Binary property Hex_Digit

pub const Hyphen: Self

Binary property Hyphen

pub const IdContinue: Self

Binary property ID_Continue

pub const IdStart: Self

Binary property ID_Start

pub const Ideographic: Self

Binary property Ideographic

pub const IdsBinaryOperator: Self

Binary property IDS_Binary_Operator

pub const IdsTrinaryOperator: Self

Binary property IDS_Trinary_Operator

pub const JoinControl: Self

Binary property Join_Control

pub const LogicalOrderException: Self

Binary property Logical_Order_Exception

pub const Lowercase: Self

Binary property Lowercase

pub const Math: Self

Binary property Math

pub const NoncharacterCodePoint: Self

Binary property Noncharacter_Code_Point

pub const QuotationMark: Self

Binary property Quotation_Mark

pub const Radical: Self

Binary property Radical

pub const SoftDotted: Self

Binary property Soft_Dotted

pub const TerminalPunctuation: Self

Binary property Terminal_Punctuation

pub const UnifiedIdeograph: Self

Binary property Unified_Ideograph

pub const Uppercase: Self

Binary property Uppercase

pub const WhiteSpace: Self

Binary property White_Space

pub const XidContinue: Self

Binary property XID_Continue

pub const XidStart: Self

Binary property XID_Start

pub const CaseSensitive: Self

Binary property Case_Sensitive

pub const SentenceTerminal: Self

Binary property Sentence_Terminal

pub const VariationSelector: Self

Binary property Variation_Selector

pub const NfdInert: Self

Binary property NFD_Inert

pub const NfkdInert: Self

Binary property NFKD_Inert

pub const NfcInert: Self

Binary property NFC_Inert

pub const NfkcInert: Self

Binary property NFKC_Inert

pub const SegmentStarter: Self

Binary property Segment_Starter

pub const PatternSyntax: Self

Binary property Pattern_Syntax

pub const PatternWhiteSpace: Self

Binary property Pattern_White_Space

pub const Alnum: Self

Binary property alnum

pub const Blank: Self

Binary property blank

pub const Graph: Self

Binary property graph

pub const Print: Self

Binary property print

pub const XDigit: Self

Binary property xdigit

pub const Cased: Self

Binary property Cased

pub const CaseIgnorable: Self

Binary property Case_Ignorable

pub const ChangesWhenLowercased: Self

Binary property Changes_When_Lowercased

pub const ChangesWhenUppercased: Self

Binary property Changes_When_Uppercased

pub const ChangesWhenTitlecased: Self

Binary property Changes_When_Titlecased

pub const ChangesWhenCasefolded: Self

Binary property Changes_When_Casefolded

pub const ChangesWhenCasemapped: Self

Binary property Changes_When_Casemapped

pub const ChangesWhenNfkcCasefolded: Self

Binary property Changes_When_NFKC_Casefolded

pub const Emoji: Self

Binary property Emoji

pub const EmojiPresentation: Self

Binary property Emoji_Presentation

pub const EmojiModifier: Self

Binary property Emoji_Modifier

pub const EmojiModifierBase: Self

Binary property Emoji_Modifier_Base

pub const EmojiComponent: Self

Binary property Emoji_Component

pub const RegionalIndicator: Self

Binary property Regional_Indicator

pub const PrependedConcatenationMark: Self

Binary property Prepended_Concatenation_Mark

pub const ExtendedPictographic: Self

Binary property Extended_Pictographic

pub const BasicEmoji: Self

Binary property Basic_Emoji

pub const EmojiKeycapSequence: Self

Binary property Emoji_Keycap_Sequence

pub const RgiEmojiModifierSequence: Self

Binary property RGI_Emoji_Modifier_Sequence

pub const RgiEmojiFlagSequence: Self

Binary property RGI_Emoji_Flag_Sequence

pub const RgiEmojiTagSequence: Self

Binary property RGI_Emoji_Tag_Sequence

pub const RgiEmojiZWJSequence: Self

Binary property RGI_Emoji_ZWJ_Sequence

pub const RgiEmoji: Self

Binary property RGI_Emoji

const BINARY_MAX: Self = Self::RgiEmoji

pub const BidiClass: Self

Enumerated property Bidi_Class

pub const Block: Self

Enumerated property Block

pub const CombiningClass: Self

Enumerated property Canonical_Combining_Class

pub const DecompositionType: Self

Enumerated property Decomposition_Type

pub const EastAsianWidth: Self

Enumerated property East_Asian_Width

pub const GeneralCategory: Self

Enumerated property General_Category

pub const JoiningGroup: Self

Enumerated property Joining_Group

pub const JoiningType: Self

Enumerated property Joining_Type

pub const LineBreak: Self

Enumerated property Line_Break

pub const NumericType: Self

Enumerated property Numeric_Type

pub const Script: Self

Enumerated property Script

pub const HangulSyllableType: Self

Enumerated property Hangul_Syllable_Type

pub const NFDQuickCheck: Self

Enumerated property NFD_Quick_Check

pub const NFKDQuickCheck: Self

Enumerated property NFKD_Quick_Check

pub const NFCQuickCheck: Self

Enumerated property NFC_Quick_Check

pub const NFKCQuickCheck: Self

Enumerated property NFKC_Quick_Check

pub const LeadCanonicalCombiningClass: Self

Enumerated property Lead_Canonical_Combining_Class

pub const TrailCanonicalCombiningClass: Self

Enumerated property Trail_Canonical_Combining_Class

pub const GraphemeClusterBreak: Self

Enumerated property Grapheme_Cluster_Break

pub const SentenceBreak: Self

Enumerated property Sentence_Break

pub const WordBreak: Self

Enumerated property Word_Break

pub const BidiPairedBracketType: Self

Enumerated property Bidi_Paired_Bracket_Type

pub const IndicPositionalCategory: Self

Enumerated property Indic_Positional_Category

pub const IndicSyllabicCategory: Self

Enumerated property Indic_Syllabic_Category

pub const VerticalOrientation: Self

Enumerated property Vertical_Orientation

const ENUMERATED_MAX: Self = Self::VerticalOrientation

pub const GeneralCategoryMask: Self

Mask property General_Category_Mask

pub const NumericValue: Self

Double property Numeric_Value

pub const Age: Self

String property Age

pub const BidiMirroringGlyph: Self

String property Bidi_Mirroring_Glyph

pub const CaseFolding: Self

String property Case_Folding

pub const ISOComment: Self

String property ISO_Comment

pub const LowercaseMapping: Self

String property Lowercase_Mapping

pub const Name: Self

String property Name

pub const SimpleCaseFolding: Self

String property Simple_Case_Folding

pub const SimpleLowercaseMapping: Self

String property Simple_Lowercase_Mapping

pub const SimpleTitlecaseMapping: Self

String property Simple_Titlecase_Mapping

pub const SimpleUppercaseMapping: Self

String property Simple_Uppercase_Mapping

pub const TitlecaseMapping: Self

String property Titlecase_Mapping

pub const Unicode1_Name: Self

String property Unicode_1_Name

pub const UppercaseMapping: Self

String property Uppercase_Mapping

pub const BidiPairedBracket: Self

String property Bidi_Paired_Bracket

const STRING_MAX: Self = Self::BidiPairedBracket

pub const ScriptExtensions: Self

Misc property Script_Extensions

impl UnicodeProperty

pub fn parse_ecma262_name(name: &str) -> Option<Self>

Given a property name (long, short, or alias), returns the corresponding UnicodeProperty value for it provided it belongs to the subset relevant for ECMA262 regexes

Returns none if the name does not match any of the names in this subset. Performs strict matching of names.

If using this to implement an ECMA262-compliant regex engine, please note these caveats:

This only returns binary and enumerated properties, as well as Self::ScriptExtensions. Lookup can be performed sufficiently with [Self::load_ecma262_binary_property_unstable()], maps::load_general_category(), maps::load_script() and script::load_script_with_extensions_unstable().

This does not handle the Any, Assigned, or ASCII pseudoproperties, since they are not defined as properties.

Any can be expressed as the range [\u{0}-\u{10FFFF}]

Assigned can be expressed as the inverse of the set gc=Cn (i.e., \P{gc=Cn}).

ASCII can be expressed as the range [\u{0}-\u{7F}]

ECMA262 regexes transparently allow General_Category_Mask values for GeneralCategory. This method does not return Self::GeneralCategoryMask, and instead relies on the caller to use mask-related lookup functions where necessary.

ECMA262 regexes allow treating General_Category (and gcm) values as binary properties, e.g. you can do things like \p{Lu} as shortform for \p{gc=Lu}. This method does not do so since these are property values, not properties, but you can use GeneralCategory::get_name_to_enum_mapper() or GeneralCategoryGroup::get_name_to_enum_mapper() to handle this.