Function icu_properties::sets::get_extender

pub fn get_extender(
    provider: &impl ResourceProvider<ExtenderV1Marker> + ?Sized
) -> UnisetResult<ExtenderV1Marker>

Characters whose principal function is to extend the value of a preceding alphabetic character or to extend the shape of adjacent characters.

Example

use icu_properties::sets;

let provider = icu_testdata::get_provider();
let payload =
    sets::get_extender(&provider)
        .expect("The data should be valid");
let data_struct = payload.get();
let extender = &data_struct.inv_list;

assert!(extender.contains('ヾ'));  // U+30FE KATAKANA VOICED ITERATION MARK
assert!(extender.contains('ー'));  // U+30FC KATAKANA-HIRAGANA PROLONGED SOUND MARK
assert!(!extender.contains('・'));  // U+30FB KATAKANA MIDDLE DOT