icu_properties::sets

Function load_alphabetic

pub fn load_alphabetic(
    provider: &(impl DataProvider<AlphabeticV1Marker> + ?Sized),
) -> Result<CodePointSetData, PropertiesError>

A version of alphabetic() that uses custom data provided by a DataProvider.

Note that this will return an owned version of the data. Functionality is available on the borrowed version, accessible through CodePointSetData::as_borrowed.