Function icu_properties::sets::load_case_sensitive

pub fn load_case_sensitive(
    provider: &(impl DataProvider<CaseSensitiveV1Marker> + ?Sized),
) -> Result<CodePointSetData, PropertiesError>

A version of case_sensitive() that uses custom data provided by a DataProvider.

Note that this will return an owned version of the data. Functionality is available on the borrowed version, accessible through CodePointSetData::as_borrowed.