Function icu_properties::sets::load_logical_order_exception

pub fn load_logical_order_exception(
    provider: &impl DataProvider<LogicalOrderExceptionV1Marker> + ?Sized
) -> Result<CodePointSetData, PropertiesError>

A small number of spacing vowel letters occurring in certain Southeast Asian scripts such as Thai and Lao

Example

use icu_properties::sets;

let data =
    sets::load_logical_order_exception(&icu_testdata::unstable())
        .expect("The data should be valid");
let logical_order_exception = data.as_borrowed();

assert!(logical_order_exception.contains('ແ'));  // U+0EC1 LAO VOWEL SIGN EI
assert!(!logical_order_exception.contains('ະ'));  // U+0EB0 LAO VOWEL SIGN A