icu_properties::sets

Function load_math

pub fn load_math(
    provider: &(impl DataProvider<MathV1Marker> + ?Sized),
) -> Result<CodePointSetData, PropertiesError>

A version of math() that uses custom data provided by a DataProvider.

Note that this will return an owned version of the data. Functionality is available on the borrowed version, accessible through CodePointSetData::as_borrowed.