Struct icu_properties::sets::CodePointSetData

pub struct CodePointSetData { /* private fields */ }

A wrapper around code point set data. It is returned by APIs that return Unicode property data in a set-like form, ex: a set of code points sharing the same value for a Unicode property. Access its data via the borrowed version, CodePointSetDataBorrowed.

Implementations

impl CodePointSetData

pub fn as_borrowed(&self) -> CodePointSetDataBorrowed<'_>

Construct a borrowed version of this type that can be queried.

This avoids a potential small underlying cost per API call (ex: contains()) by consolidating it up front.

use icu_properties::sets;

let data = sets::load_alphabetic(&icu_testdata::unstable())
    .expect("The data should be valid");

let alphabetic = data.as_borrowed();

assert!(!alphabetic.contains('3'));
assert!(alphabetic.contains('A'));

pub fn from_data<M>(data: DataPayload<M>) -> Selfwhere M: DataMarker<Yokeable = PropertyCodePointSetV1<'static>>,

Construct a new one from loaded data

Typically it is preferable to use getters like load_ascii_hex_digit() instead

pub fn from_code_point_inversion_list( set: CodePointInversionList<'static> ) -> Self

Construct a new owned CodePointInversionList

pub fn as_code_point_inversion_list( &self ) -> Option<&CodePointInversionList<'_>>

Convert this type to a CodePointInversionList as a borrowed value.

The data backing this is extensible and supports multiple implementations. Currently it is always CodePointInversionList; however in the future more backends may be added, and users may select which at data generation time.

This method returns an Option in order to return None when the backing data provider cannot return a CodePointInversionList, or cannot do so within the expected constant time constraint.

pub fn to_code_point_inversion_list(&self) -> CodePointInversionList<'_>

Convert this type to a CodePointInversionList, borrowing if possible, otherwise allocating a new CodePointInversionList.

The data backing this is extensible and supports multiple implementations. Currently it is always CodePointInversionList; however in the future more backends may be added, and users may select which at data generation time.

The performance of the conversion to this specific return type will vary depending on the data structure that is backing self.