icu_properties

Struct CodePointMapData

pub struct CodePointMapData<T: TrieValue> { /* private fields */ }

A wrapper around code point map data.

It is returned by APIs that return Unicode property data in a map-like form, ex: enumerated property value data keyed by code point. Access its data via the borrowed version, CodePointMapDataBorrowed.

Implementations

impl<T: TrieValue> CodePointMapData<T>

pub const fn new() -> CodePointMapDataBorrowed<'static, T>
where T: EnumeratedProperty,

Creates a new CodePointMapData for a EnumeratedProperty.

See the documentation on EnumeratedProperty implementations for details.

โจ Enabled with the compiled_data Cargo feature.

๐ Help choosing a constructor

pub fn try_new_unstable( provider: &(impl DataProvider<T::DataMarker> + ?Sized), ) -> Result<Self, DataError>
where T: EnumeratedProperty,

A version of Self::new that uses custom data provided by a DataProvider.

๐ Help choosing a constructor

โ๏ธ The bounds on provider may change over time, including in SemVer minor releases.

pub fn as_borrowed(&self) -> CodePointMapDataBorrowed<'_, T>

Construct a borrowed version of this type that can be queried.

This avoids a potential small underlying cost per API call (like get()) by consolidating it up front.

This owned version if returned by functions that use a runtime data provider.

pub fn from_code_point_trie(trie: CodePointTrie<'static, T>) -> Self

Construct a new one an owned CodePointTrie

pub fn as_code_point_trie(&self) -> Option<&CodePointTrie<'_, T>>

Convert this type to a CodePointTrie as a borrowed value.

The data backing this is extensible and supports multiple implementations. Currently it is always CodePointTrie; however in the future more backends may be added, and users may select which at data generation time.

This method returns an Option in order to return None when the backing data provider cannot return a CodePointTrie, or cannot do so within the expected constant time constraint.

pub fn to_code_point_trie(&self) -> CodePointTrie<'_, T>

Convert this type to a CodePointTrie, borrowing if possible, otherwise allocating a new CodePointTrie.

The data backing this is extensible and supports multiple implementations. Currently it is always CodePointTrie; however in the future more backends may be added, and users may select which at data generation time.

The performance of the conversion to this specific return type will vary depending on the data structure that is backing self.