icu_properties

Struct CodePointMapDataBorrowed

pub struct CodePointMapDataBorrowed<'a, T: TrieValue> { /* private fields */ }

A borrowed wrapper around code point set data, returned by [CodePointSetData::as_borrowed()]. More efficient to query.

Implementations

impl<'a, T: TrieValue> CodePointMapDataBorrowed<'a, T>

pub fn get(self, ch: char) -> T

Get the value this map has associated with code point ch

Example

use icu::properties::CodePointMapData;
use icu::properties::props::GeneralCategory;

let gc = CodePointMapData::<GeneralCategory>::new();

assert_eq!(gc.get('ๆจ'), GeneralCategory::OtherLetter);  // U+6728
assert_eq!(gc.get('๐'), GeneralCategory::OtherSymbol);  // U+1F383 JACK-O-LANTERN

pub fn get32(self, ch: u32) -> T

See Self::get.

pub fn iter_ranges(self) -> impl Iterator<Item = CodePointMapRange<T>> + 'a

Yields an Iterator returning ranges of consecutive code points that share the same value in the CodePointMapData.

Examples

use icu::properties::props::GeneralCategory;
use icu::properties::CodePointMapData;

let gc = CodePointMapData::<GeneralCategory>::new();
let mut ranges = gc.iter_ranges();
let next = ranges.next().unwrap();
assert_eq!(next.range, 0..=31);
assert_eq!(next.value, GeneralCategory::Control);
let next = ranges.next().unwrap();
assert_eq!(next.range, 32..=32);
assert_eq!(next.value, GeneralCategory::SpaceSeparator);

pub fn iter_ranges_for_value( self, val: T, ) -> impl Iterator<Item = RangeInclusive<u32>> + 'a

Yields an Iterator returning ranges of consecutive code points that share the same value v in the CodePointMapData.

Examples

use icu::properties::props::GeneralCategory;
use icu::properties::CodePointMapData;

let gc = CodePointMapData::<GeneralCategory>::new();
let mut ranges = gc.iter_ranges_for_value(GeneralCategory::UppercaseLetter);
assert_eq!(ranges.next().unwrap(), 'A' as u32..='Z' as u32);
assert_eq!(ranges.next().unwrap(), 'ร' as u32..='ร' as u32);
assert_eq!(ranges.next().unwrap(), 'ร' as u32..='ร' as u32);

pub fn iter_ranges_for_value_complemented( self, val: T, ) -> impl Iterator<Item = RangeInclusive<u32>> + 'a

Yields an Iterator returning ranges of consecutive code points that do not have the value v in the CodePointMapData.

impl<T: TrieValue> CodePointMapDataBorrowed<'static, T>

pub const fn new() -> Self
where T: EnumeratedProperty,

Creates a new CodePointMapDataBorrowed for a EnumeratedProperty.

See the documentation on EnumeratedProperty implementations for details.

โจ Enabled with the compiled_data Cargo feature.

๐ Help choosing a constructor

pub const fn static_to_owned(self) -> CodePointMapData<T>

Cheaply converts a CodePointMapDataBorrowed<'static> into a CodePointMapData.

Note: Due to branching and indirection, using CodePointMapData might inhibit some compile-time optimizations that are possible with CodePointMapDataBorrowed.

impl<'a> CodePointMapDataBorrowed<'a, GeneralCategory>

pub fn iter_ranges_for_group( self, group: GeneralCategoryGroup, ) -> impl Iterator<Item = RangeInclusive<u32>> + 'a

Yields an Iterator returning ranges of consecutive code points that have a General_Category value belonging to the specified GeneralCategoryGroup

Examples

use icu::properties::props::{GeneralCategory, GeneralCategoryGroup};
use icu::properties::CodePointMapData;

let gc = CodePointMapData::<GeneralCategory>::new();
let mut ranges = gc.iter_ranges_for_group(GeneralCategoryGroup::Letter);
assert_eq!(ranges.next().unwrap(), 'A' as u32..='Z' as u32);
assert_eq!(ranges.next().unwrap(), 'a' as u32..='z' as u32);
assert_eq!(ranges.next().unwrap(), 'ยช' as u32..='ยช' as u32);
assert_eq!(ranges.next().unwrap(), 'ยต' as u32..='ยต' as u32);
assert_eq!(ranges.next().unwrap(), 'ยบ' as u32..='ยบ' as u32);
assert_eq!(ranges.next().unwrap(), 'ร' as u32..='ร' as u32);
assert_eq!(ranges.next().unwrap(), 'ร' as u32..='รถ' as u32);