icu_properties

Struct EastAsianWidth

#[repr(transparent)]
pub struct EastAsianWidth(pub u8);

Enumerated property East_Asian_Width.

See โDefinitionโ in UAX #11 for the summary of each property value: https://www.unicode.org/reports/tr11/#Definitions

The numeric value is compatible with UEastAsianWidth in ICU4C.

Tuple Fields
0: u8
Implementations

impl EastAsianWidth

pub const Neutral: EastAsianWidth

pub const Ambiguous: EastAsianWidth

pub const Halfwidth: EastAsianWidth

pub const Fullwidth: EastAsianWidth

pub const Narrow: EastAsianWidth

pub const Wide: EastAsianWidth

impl EastAsianWidth

pub const fn name_to_enum_mapper() -> PropertyValueNameToEnumMapperBorrowed<'static, EastAsianWidth>

Return a PropertyValueNameToEnumMapper, capable of looking up values from strings for the East_Asian_Width enumerated property.

โจ Enabled with the compiled_data Cargo feature.

๐ Help choosing a constructor

Example

use icu::properties::EastAsianWidth;

let lookup = EastAsianWidth::name_to_enum_mapper();
// short name for value
assert_eq!(lookup.get_strict("N"), Some(EastAsianWidth::Neutral));
assert_eq!(lookup.get_strict("H"), Some(EastAsianWidth::Halfwidth));
// long name for value
assert_eq!(lookup.get_strict("Neutral"), Some(EastAsianWidth::Neutral));
assert_eq!(lookup.get_strict("Halfwidth"), Some(EastAsianWidth::Halfwidth));
// name has incorrect casing / extra hyphen
assert_eq!(lookup.get_strict("half-width"), None);
// loose matching of name
assert_eq!(lookup.get_loose("half-width"), Some(EastAsianWidth::Halfwidth));
// fake property
assert_eq!(lookup.get_strict("TwoPointFiveWidth"), None);

pub fn get_name_to_enum_mapper( provider: &(impl DataProvider<EastAsianWidthNameToValueV1Marker> + ?Sized), ) -> Result<PropertyValueNameToEnumMapper<EastAsianWidth>, PropertiesError>

A version of EastAsianWidth::name_to_enum_mapper() that uses custom data provided by a DataProvider.

๐ Help choosing a constructor

pub const fn enum_to_short_name_mapper() -> PropertyEnumToValueNameLinearMapperBorrowed<'static, EastAsianWidth>

Return a PropertyEnumToValueNameLinearMapper, capable of looking up short names for values of the East_Asian_Width enumerated property.

โจ Enabled with the compiled_data Cargo feature.

๐ Help choosing a constructor

Example

use icu::properties::EastAsianWidth;

let lookup = EastAsianWidth::enum_to_short_name_mapper();
assert_eq!(lookup.get(EastAsianWidth::Neutral), Some("N"));
assert_eq!(lookup.get(EastAsianWidth::Halfwidth), Some("H"));

pub fn get_enum_to_short_name_mapper( provider: &(impl DataProvider<EastAsianWidthValueToShortNameV1Marker> + ?Sized), ) -> Result<PropertyEnumToValueNameLinearMapper<EastAsianWidth>, PropertiesError>

A version of EastAsianWidth::enum_to_short_name_mapper() that uses custom data provided by a DataProvider.

๐ Help choosing a constructor

pub const fn enum_to_long_name_mapper() -> PropertyEnumToValueNameLinearMapperBorrowed<'static, EastAsianWidth>

Return a PropertyEnumToValueNameLinearMapper, capable of looking up long names for values of the East_Asian_Width enumerated property.

โจ Enabled with the compiled_data Cargo feature.

๐ Help choosing a constructor

Example

use icu::properties::EastAsianWidth;

let lookup = EastAsianWidth::enum_to_long_name_mapper();
assert_eq!(lookup.get(EastAsianWidth::Neutral), Some("Neutral"));
assert_eq!(lookup.get(EastAsianWidth::Halfwidth), Some("Halfwidth"));

pub fn get_enum_to_long_name_mapper( provider: &(impl DataProvider<EastAsianWidthValueToLongNameV1Marker> + ?Sized), ) -> Result<PropertyEnumToValueNameLinearMapper<EastAsianWidth>, PropertiesError>

A version of EastAsianWidth::enum_to_long_name_mapper() that uses custom data provided by a DataProvider.

๐ Help choosing a constructor