Struct icu_properties::GraphemeClusterBreak

#[repr(transparent)]
pub struct GraphemeClusterBreak(pub u8);

Enumerated property Grapheme_Cluster_Break.

See “Default Grapheme Cluster Boundary Specification” in UAX #29 for the summary of each property value: https://www.unicode.org/reports/tr29/#Default_Grapheme_Cluster_Table

The numeric value is compatible with UGraphemeClusterBreak in ICU4C.

Tuple Fields
0: u8
Implementations

impl GraphemeClusterBreak

pub const Other: GraphemeClusterBreak = _

pub const Control: GraphemeClusterBreak = _

pub const CR: GraphemeClusterBreak = _

pub const Extend: GraphemeClusterBreak = _

pub const L: GraphemeClusterBreak = _

pub const LF: GraphemeClusterBreak = _

pub const LV: GraphemeClusterBreak = _

pub const LVT: GraphemeClusterBreak = _

pub const T: GraphemeClusterBreak = _

pub const V: GraphemeClusterBreak = _

pub const SpacingMark: GraphemeClusterBreak = _

pub const Prepend: GraphemeClusterBreak = _

pub const RegionalIndicator: GraphemeClusterBreak = _

pub const EBase: GraphemeClusterBreak = _

This value is obsolete and unused.

pub const EBaseGAZ: GraphemeClusterBreak = _

This value is obsolete and unused.

pub const EModifier: GraphemeClusterBreak = _

This value is obsolete and unused.

pub const GlueAfterZwj: GraphemeClusterBreak = _

This value is obsolete and unused.

pub const ZWJ: GraphemeClusterBreak = _

impl GraphemeClusterBreak

pub fn get_name_to_enum_mapper( provider: &impl DataProvider<GraphemeClusterBreakNameToValueV1Marker> + ?Sized ) -> Result<PropertyValueNameToEnumMapper<GraphemeClusterBreak>, PropertiesError>

Return a PropertyValueNameToEnumMapper, capable of looking up values from strings for the Grapheme_Cluster_Break enumerated property

Example

use icu::properties::GraphemeClusterBreak;

let lookup = GraphemeClusterBreak::get_name_to_enum_mapper(&icu_testdata::unstable())
                 .expect("The data should be valid");
let lookup = lookup.as_borrowed();
// short name for value
assert_eq!(lookup.get_strict("EX"), Some(GraphemeClusterBreak::Extend));
assert_eq!(lookup.get_strict("RI"), Some(GraphemeClusterBreak::RegionalIndicator));
// long name for value
assert_eq!(lookup.get_strict("Extend"), Some(GraphemeClusterBreak::Extend));
assert_eq!(lookup.get_strict("Regional_Indicator"), Some(GraphemeClusterBreak::RegionalIndicator));
// name has incorrect casing and lacks an underscore
assert_eq!(lookup.get_strict("regionalindicator"), None);
// loose matching of name
assert_eq!(lookup.get_loose("regionalindicator"), Some(GraphemeClusterBreak::RegionalIndicator));
// fake property
assert_eq!(lookup.get_strict("Regional_Indicator_Two_Point_Oh"), None);

pub fn get_enum_to_short_name_mapper( provider: &impl DataProvider<GraphemeClusterBreakValueToShortNameV1Marker> + ?Sized ) -> Result<PropertyEnumToValueNameLinearMapper<GraphemeClusterBreak>, PropertiesError>

Return a PropertyEnumToValueNameLinearMapper, capable of looking up short names for values of the Grapheme_Cluster_Break enumerated property

Example

use icu::properties::GraphemeClusterBreak;

let lookup = GraphemeClusterBreak::get_enum_to_short_name_mapper(&icu_testdata::unstable())
                 .expect("The data should be valid");
let lookup = lookup.as_borrowed();
assert_eq!(lookup.get(GraphemeClusterBreak::Extend), Some("EX"));
assert_eq!(lookup.get(GraphemeClusterBreak::RegionalIndicator), Some("RI"));

pub fn get_enum_to_long_name_mapper( provider: &impl DataProvider<GraphemeClusterBreakValueToLongNameV1Marker> + ?Sized ) -> Result<PropertyEnumToValueNameLinearMapper<GraphemeClusterBreak>, PropertiesError>

Return a PropertyEnumToValueNameLinearMapper, capable of looking up long names for values of the Grapheme_Cluster_Break enumerated property

Example

use icu::properties::GraphemeClusterBreak;

let lookup = GraphemeClusterBreak::get_enum_to_long_name_mapper(&icu_testdata::unstable())
                 .expect("The data should be valid");
let lookup = lookup.as_borrowed();
assert_eq!(lookup.get(GraphemeClusterBreak::Extend), Some("Extend"));
assert_eq!(lookup.get(GraphemeClusterBreak::RegionalIndicator), Some("Regional_Indicator"));