Struct icu_properties::IndicSyllabicCategory

#[repr(transparent)]
pub struct IndicSyllabicCategory(pub u8);

Property Indic_Syllabic_Category. See UAX #44: https://www.unicode.org/reports/tr44/#Indic_Syllabic_Category.

The numeric value is compatible with UIndicSyllabicCategory in ICU4C.

Tuple Fields
0: u8
Implementations

impl IndicSyllabicCategory

pub const Other: IndicSyllabicCategory = _

pub const Avagraha: IndicSyllabicCategory = _

pub const Bindu: IndicSyllabicCategory = _

pub const BrahmiJoiningNumber: IndicSyllabicCategory = _

pub const CantillationMark: IndicSyllabicCategory = _

pub const Consonant: IndicSyllabicCategory = _

pub const ConsonantDead: IndicSyllabicCategory = _

pub const ConsonantFinal: IndicSyllabicCategory = _

pub const ConsonantHeadLetter: IndicSyllabicCategory = _

pub const ConsonantInitialPostfixed: IndicSyllabicCategory = _

pub const ConsonantKiller: IndicSyllabicCategory = _

pub const ConsonantMedial: IndicSyllabicCategory = _

pub const ConsonantPlaceholder: IndicSyllabicCategory = _

pub const ConsonantPrecedingRepha: IndicSyllabicCategory = _

pub const ConsonantPrefixed: IndicSyllabicCategory = _

pub const ConsonantSucceedingRepha: IndicSyllabicCategory = _

pub const ConsonantSubjoined: IndicSyllabicCategory = _

pub const ConsonantWithStacker: IndicSyllabicCategory = _

pub const GeminationMark: IndicSyllabicCategory = _

pub const InvisibleStacker: IndicSyllabicCategory = _

pub const Joiner: IndicSyllabicCategory = _

pub const ModifyingLetter: IndicSyllabicCategory = _

pub const NonJoiner: IndicSyllabicCategory = _

pub const Nukta: IndicSyllabicCategory = _

pub const Number: IndicSyllabicCategory = _

pub const NumberJoiner: IndicSyllabicCategory = _

pub const PureKiller: IndicSyllabicCategory = _

pub const RegisterShifter: IndicSyllabicCategory = _

pub const SyllableModifier: IndicSyllabicCategory = _

pub const ToneLetter: IndicSyllabicCategory = _

pub const ToneMark: IndicSyllabicCategory = _

pub const Virama: IndicSyllabicCategory = _

pub const Visarga: IndicSyllabicCategory = _

pub const Vowel: IndicSyllabicCategory = _

pub const VowelDependent: IndicSyllabicCategory = _

pub const VowelIndependent: IndicSyllabicCategory = _

impl IndicSyllabicCategory

pub const fn name_to_enum_mapper() -> PropertyValueNameToEnumMapperBorrowed<'static, IndicSyllabicCategory>

Return a PropertyValueNameToEnumMapper, capable of looking up values from strings for the Indic_Syllabic_Category enumerated property.

✨ Enabled with the compiled_data Cargo feature.

📚 Help choosing a constructor

Example

use icu::properties::IndicSyllabicCategory;

let lookup = IndicSyllabicCategory::name_to_enum_mapper();
// long/short name for value
assert_eq!(lookup.get_strict("Brahmi_Joining_Number"), Some(IndicSyllabicCategory::BrahmiJoiningNumber));
assert_eq!(lookup.get_strict("Vowel_Independent"), Some(IndicSyllabicCategory::VowelIndependent));
// name has incorrect casing and hyphens
assert_eq!(lookup.get_strict("brahmi-joining-number"), None);
// loose matching of name
assert_eq!(lookup.get_loose("brahmi-joining-number"), Some(IndicSyllabicCategory::BrahmiJoiningNumber));
// fake property
assert_eq!(lookup.get_strict("Tone_Number"), None);

pub fn get_name_to_enum_mapper( provider: &(impl DataProvider<IndicSyllabicCategoryNameToValueV1Marker> + ?Sized), ) -> Result<PropertyValueNameToEnumMapper<IndicSyllabicCategory>, PropertiesError>

A version of IndicSyllabicCategory::name_to_enum_mapper() that uses custom data provided by a DataProvider.

📚 Help choosing a constructor

pub const fn enum_to_short_name_mapper() -> PropertyEnumToValueNameLinearMapperBorrowed<'static, IndicSyllabicCategory>

Return a PropertyEnumToValueNameLinearMapper, capable of looking up short names for values of the Indic_Syllabic_Category enumerated property.

✨ Enabled with the compiled_data Cargo feature.

📚 Help choosing a constructor

Example

use icu::properties::IndicSyllabicCategory;

let lookup = IndicSyllabicCategory::enum_to_short_name_mapper();
assert_eq!(lookup.get(IndicSyllabicCategory::BrahmiJoiningNumber), Some("Brahmi_Joining_Number"));
assert_eq!(lookup.get(IndicSyllabicCategory::VowelIndependent), Some("Vowel_Independent"));

pub fn get_enum_to_short_name_mapper( provider: &(impl DataProvider<IndicSyllabicCategoryValueToShortNameV1Marker> + ?Sized), ) -> Result<PropertyEnumToValueNameLinearMapper<IndicSyllabicCategory>, PropertiesError>

A version of IndicSyllabicCategory::enum_to_short_name_mapper() that uses custom data provided by a DataProvider.

📚 Help choosing a constructor

pub const fn enum_to_long_name_mapper() -> PropertyEnumToValueNameLinearMapperBorrowed<'static, IndicSyllabicCategory>

Return a PropertyEnumToValueNameLinearMapper, capable of looking up long names for values of the Indic_Syllabic_Category enumerated property.

✨ Enabled with the compiled_data Cargo feature.

📚 Help choosing a constructor

Example

use icu::properties::IndicSyllabicCategory;

let lookup = IndicSyllabicCategory::enum_to_long_name_mapper();
assert_eq!(lookup.get(IndicSyllabicCategory::BrahmiJoiningNumber), Some("Brahmi_Joining_Number"));
assert_eq!(lookup.get(IndicSyllabicCategory::VowelIndependent), Some("Vowel_Independent"));

pub fn get_enum_to_long_name_mapper( provider: &(impl DataProvider<IndicSyllabicCategoryValueToLongNameV1Marker> + ?Sized), ) -> Result<PropertyEnumToValueNameLinearMapper<IndicSyllabicCategory>, PropertiesError>

A version of IndicSyllabicCategory::enum_to_long_name_mapper() that uses custom data provided by a DataProvider.

📚 Help choosing a constructor