icu_properties

Struct PropertyNamesShortBorrowed

pub struct PropertyNamesShortBorrowed<'a, T: NamedEnumeratedProperty> { /* private fields */ }

A borrowed wrapper around property value name-to-enum data, returned by PropertyNamesShort::as_borrowed(). More efficient to query.

Implementations

impl<'a, T: NamedEnumeratedProperty> PropertyNamesShortBorrowed<'a, T>

pub fn get(self, property: T) -> Option<&'a str>

Get the property name given a value

Example

use icu::properties::props::CanonicalCombiningClass;
use icu::properties::PropertyNamesShort;

let lookup = PropertyNamesShort::<CanonicalCombiningClass>::new();
assert_eq!(lookup.get(CanonicalCombiningClass::KanaVoicing), Some("KV"));
assert_eq!(lookup.get(CanonicalCombiningClass::AboveLeft), Some("AL"));

impl PropertyNamesShortBorrowed<'_, Script>

pub fn get_locale_script(self, property: Script) -> Option<Script>

Gets the โnameโ of a script property as a icu::locale::subtags::Script.

This method is available only on PropertyNamesShortBorrowed<Script>.

Example

use icu::locale::subtags::script;
use icu::properties::props::Script;
use icu::properties::PropertyNamesShort;

let lookup = PropertyNamesShort::<Script>::new();
assert_eq!(
    lookup.get_locale_script(Script::Brahmi),
    Some(script!("Brah"))
);
assert_eq!(
    lookup.get_locale_script(Script::Hangul),
    Some(script!("Hang"))
);

For the reverse direction, use property parsing as normal:

use icu::locale::subtags::script;
use icu::properties::props::Script;
use icu::properties::PropertyParser;

let parser = PropertyParser::<Script>::new();
assert_eq!(
    parser.get_strict(script!("Brah").as_str()),
    Some(Script::Brahmi)
);
assert_eq!(
    parser.get_strict(script!("Hang").as_str()),
    Some(Script::Hangul)
);

impl<T: NamedEnumeratedProperty> PropertyNamesShortBorrowed<'static, T>

pub fn new() -> Self

Creates a new instance of PropertyNamesShortBorrowed<T>.

โจ Enabled with the compiled_data Cargo feature.

๐ Help choosing a constructor

pub fn static_to_owned(self) -> PropertyNamesShort<T>

Cheaply converts a PropertyNamesShortBorrowed<'static> into a PropertyNamesShort.

Note: Due to branching and indirection, using PropertyNamesShort might inhibit some compile-time optimizations that are possible with PropertyNamesShortBorrowed.

This is currently not const unlike other static_to_owned() functions since it needs const traits to do that safely