icu_properties

Struct PropertyParser

pub struct PropertyParser<T> { /* private fields */ }

A struct capable of looking up a property value from a string name. Access its data by calling Self::as_borrowed() and using the methods on PropertyParserBorrowed.

The name can be a short name (Lu), a long name(Uppercase_Letter), or an alias.

Property names can be looked up using โstrictโ matching (looking for a name that matches exactly), or โloose matchingโ, where the name is allowed to deviate in terms of ASCII casing, whitespace, underscores, and hyphens.

Example

use icu::properties::props::GeneralCategory;
use icu::properties::PropertyParser;

let lookup = PropertyParser::<GeneralCategory>::new();
// short name for value
assert_eq!(
    lookup.get_strict("Lu"),
    Some(GeneralCategory::UppercaseLetter)
);
assert_eq!(
    lookup.get_strict("Pd"),
    Some(GeneralCategory::DashPunctuation)
);
// long name for value
assert_eq!(
    lookup.get_strict("Uppercase_Letter"),
    Some(GeneralCategory::UppercaseLetter)
);
assert_eq!(
    lookup.get_strict("Dash_Punctuation"),
    Some(GeneralCategory::DashPunctuation)
);
// name has incorrect casing
assert_eq!(lookup.get_strict("dashpunctuation"), None);
// loose matching of name
assert_eq!(
    lookup.get_loose("dash-punctuation"),
    Some(GeneralCategory::DashPunctuation)
);
// fake property
assert_eq!(lookup.get_strict("Animated_Gif"), None);

Implementations

impl<T> PropertyParser<T>

pub fn new() -> PropertyParserBorrowed<'static, T>
where T: ParseableEnumeratedProperty,

Creates a new instance of PropertyParser<T> using compiled data.

โจ Enabled with the compiled_data Cargo feature.

๐ Help choosing a constructor

pub fn try_new_unstable( provider: &(impl DataProvider<T::DataMarker> + ?Sized), ) -> Result<Self, DataError>
where T: ParseableEnumeratedProperty,

A version of Self::new that uses custom data provided by a DataProvider.

๐ Help choosing a constructor

โ๏ธ The bounds on provider may change over time, including in SemVer minor releases.

pub fn as_borrowed(&self) -> PropertyParserBorrowed<'_, T>

Construct a borrowed version of this type that can be queried.

This avoids a potential small underlying cost per API call (like get_strict()) by consolidating it up front.