Struct icu_properties::WordBreak

#[repr(transparent)]
pub struct WordBreak(pub u8);

Enumerated property Word_Break.

See “Default Word Boundary Specification” in UAX #29 for the summary of each property value: https://www.unicode.org/reports/tr29/#Default_Word_Boundaries.

The numeric value is compatible with UWordBreakValues in ICU4C.

Tuple Fields
0: u8
Implementations

impl WordBreak

pub const Other: WordBreak = _

pub const ALetter: WordBreak = _

pub const Format: WordBreak = _

pub const Katakana: WordBreak = _

pub const MidLetter: WordBreak = _

pub const MidNum: WordBreak = _

pub const Numeric: WordBreak = _

pub const ExtendNumLet: WordBreak = _

pub const CR: WordBreak = _

pub const Extend: WordBreak = _

pub const LF: WordBreak = _

pub const MidNumLet: WordBreak = _

pub const Newline: WordBreak = _

pub const RegionalIndicator: WordBreak = _

pub const HebrewLetter: WordBreak = _

pub const SingleQuote: WordBreak = _

pub const DoubleQuote: WordBreak = _

pub const EBase: WordBreak = _

This value is obsolete and unused.

pub const EBaseGAZ: WordBreak = _

This value is obsolete and unused.

pub const EModifier: WordBreak = _

This value is obsolete and unused.

pub const GlueAfterZwj: WordBreak = _

This value is obsolete and unused.

pub const ZWJ: WordBreak = _

pub const WSegSpace: WordBreak = _

impl WordBreak

pub fn get_name_to_enum_mapper( provider: &impl DataProvider<WordBreakNameToValueV1Marker> + ?Sized ) -> Result<PropertyValueNameToEnumMapper<WordBreak>, PropertiesError>

Return a PropertyValueNameToEnumMapper, capable of looking up values from strings for the Word_Break enumerated property

Example

use icu::properties::WordBreak;

let lookup = WordBreak::get_name_to_enum_mapper(&icu_testdata::unstable())
                 .expect("The data should be valid");
let lookup = lookup.as_borrowed();
// short name for value
assert_eq!(lookup.get_strict("KA"), Some(WordBreak::Katakana));
assert_eq!(lookup.get_strict("LE"), Some(WordBreak::ALetter));
// long name for value
assert_eq!(lookup.get_strict("Katakana"), Some(WordBreak::Katakana));
assert_eq!(lookup.get_strict("ALetter"), Some(WordBreak::ALetter));
// name has incorrect casing
assert_eq!(lookup.get_strict("Aletter"), None);
// loose matching of name
assert_eq!(lookup.get_loose("Aletter"), Some(WordBreak::ALetter));
// fake property
assert_eq!(lookup.get_strict("Quadruple_Quote"), None);

pub fn get_enum_to_short_name_mapper( provider: &impl DataProvider<WordBreakValueToShortNameV1Marker> + ?Sized ) -> Result<PropertyEnumToValueNameLinearMapper<WordBreak>, PropertiesError>

Return a PropertyEnumToValueNameLinearMapper, capable of looking up short names for values of the Word_Break enumerated property

Example

use icu::properties::WordBreak;

let lookup = WordBreak::get_enum_to_short_name_mapper(&icu_testdata::unstable())
                 .expect("The data should be valid");
let lookup = lookup.as_borrowed();
assert_eq!(lookup.get(WordBreak::Katakana), Some("KA"));
assert_eq!(lookup.get(WordBreak::ALetter), Some("LE"));

pub fn get_enum_to_long_name_mapper( provider: &impl DataProvider<WordBreakValueToLongNameV1Marker> + ?Sized ) -> Result<PropertyEnumToValueNameLinearMapper<WordBreak>, PropertiesError>

Return a PropertyEnumToValueNameLinearMapper, capable of looking up long names for values of the Word_Break enumerated property

Example

use icu::properties::WordBreak;

let lookup = WordBreak::get_enum_to_long_name_mapper(&icu_testdata::unstable())
                 .expect("The data should be valid");
let lookup = lookup.as_borrowed();
assert_eq!(lookup.get(WordBreak::Katakana), Some("Katakana"));
assert_eq!(lookup.get(WordBreak::ALetter), Some("ALetter"));