Crate icu_properties

Definitions of Unicode Properties and APIs for retrieving property data in an appropriate data structure.

This module is published as its own crate (icu_properties) and as part of the icu crate. See the latter for more details on the ICU4X project.

APIs that return a CodePointSetData exist for binary properties and certain enumerated properties. See the sets module for more details.

APIs that return a CodePointMapData exist for certain enumerated properties. See the maps module for more details.

Examples

Property data as CodePointSetDatas

use icu::properties::{maps, sets, GeneralCategory};

// A binary property as a `CodePointSetData`

assert!(sets::emoji().contains('🎃')); // U+1F383 JACK-O-LANTERN
assert!(!sets::emoji().contains('木')); // U+6728

// An individual enumerated property value as a `CodePointSetData`

let line_sep_data = maps::general_category()
    .get_set_for_value(GeneralCategory::LineSeparator);
let line_sep = line_sep_data.as_borrowed();

assert!(line_sep.contains32(0x2028));
assert!(!line_sep.contains32(0x2029));

Property data as CodePointMapDatas

use icu::properties::{maps, Script};

assert_eq!(maps::script().get('🎃'), Script::Common); // U+1F383 JACK-O-LANTERN
assert_eq!(maps::script().get('木'), Script::Han); // U+6728

Re-exports

pub use PropertiesError as Error;

Modules

bidi_data

Data and APIs for supporting specific Bidi properties data in an efficient structure.

exemplar_chars

This module provides APIs for getting exemplar characters for a locale.

maps

The functions in this module return a CodePointMapData representing, for each code point in the entire range of code points, the property values for a particular Unicode property.

names

Module for working with the names of property values

provider

🚧 [Unstable] Data provider struct definitions for this ICU4X component.

script

Data and APIs for supporting both Script and Script_Extensions property values in an efficient structure.

sets

The functions in this module return a CodePointSetData containing the set of characters with a particular Unicode property.

Structs

BidiClass

Enumerated property Bidi_Class

CanonicalCombiningClass

Property Canonical_Combining_Class. See UAX #15: https://www.unicode.org/reports/tr15/.

EastAsianWidth

Enumerated property East_Asian_Width.

GeneralCategoryGroup

Groupings of multiple General_Category property values.

GraphemeClusterBreak

Enumerated property Grapheme_Cluster_Break.

IndicSyllabicCategory

Property Indic_Syllabic_Category. See UAX #44: https://www.unicode.org/reports/tr44/#Indic_Syllabic_Category.

LineBreak

Enumerated property Line_Break.

Script

Enumerated property Script.

SentenceBreak

Enumerated property Sentence_Break. See “Default Sentence Boundary Specification” in UAX #29 for the summary of each property value: https://www.unicode.org/reports/tr29/#Default_Word_Boundaries.

WordBreak

Enumerated property Word_Break.

Enums

GeneralCategory

Enumerated property General_Category.

PropertiesError

A list of error outcomes for various operations in this module.