Module icu_properties::maps

The functions in this module return a CodePointMapData representing, for each code point in the entire range of code points, the property values for a particular Unicode property.

The descriptions of most properties are taken from TR44, the documentation for the Unicode Character Database.

Structs

CodePointMapData

A wrapper around code point map data. It is returned by APIs that return Unicode property data in a map-like form, ex: enumerated property value data keyed by code point. Access its data via the borrowed version, CodePointMapDataBorrowed.

CodePointMapDataBorrowed

A borrowed wrapper around code point set data, returned by CodePointSetData::as_borrowed(). More efficient to query.

Functions

load_bidi_class

Return a CodePointMapData for the Bidi_Class Unicode enumerated property. See BidiClass.

load_canonical_combining_class

Return a CodePointMapData for the Canonical_Combining_Class Unicode property. See CanonicalCombiningClass.

load_east_asian_width

Return a CodePointMapData for the East_Asian_Width Unicode enumerated property. See EastAsianWidth.

load_general_category

Return a CodePointMapData for the General_Category Unicode enumerated property. See GeneralCategory.

load_grapheme_cluster_break

Return a CodePointMapData for the Grapheme_Cluster_Break Unicode enumerated property. See GraphemeClusterBreak.

load_line_break

Return a CodePointMapData for the Line_Break Unicode enumerated property. See LineBreak.

load_script

Return a CodePointMapData for the Script Unicode enumerated property. See Script.

load_sentence_break

Return a CodePointMapData for the Sentence_Break Unicode enumerated property. See SentenceBreak.

load_word_break

Return a CodePointMapData for the Word_Break Unicode enumerated property. See WordBreak.