icu_properties

Module props

A collection of property definitions shared across contexts (ex: representing trie values).

This module defines enums / newtypes for enumerated properties. String properties are represented as newtypes if their values represent code points.

Macros

create_const_array 🔒

See [test_enumerated_property_completeness] for usage. Example input:

impl_value_getter 🔒

Structs

BidiClass

Enumerated property Bidi_Class

BidiClassULE

ULE type for BidiClass

CanonicalCombiningClass

Property Canonical_Combining_Class. See UAX #15: https://www.unicode.org/reports/tr15/.

CanonicalCombiningClassULE

ULE type for CanonicalCombiningClass

EastAsianWidth

Enumerated property East_Asian_Width.

EastAsianWidthULE

ULE type for EastAsianWidth

ErasedEnumToValueNameLinearMapV1Marker 🔒

Private marker type for PropertyEnumToValueNameLinearMapper to work for all properties at once

ErasedEnumToValueNameLinearTiny4MapV1Marker 🔒

Private marker type for PropertyEnumToValueNameLinearTiny4Mapper to work for all properties at once

ErasedEnumToValueNameSparseMapV1Marker 🔒

Private marker type for PropertyEnumToValueNameSparseMapper to work for all properties at once

ErasedNameToEnumMapV1Marker 🔒

Private marker type for PropertyValueNameToEnumMapper to work for all properties at once

GeneralCategoryGroup

Groupings of multiple General_Category property values.

GeneralCategoryTryFromError

GeneralCategoryULE

ULE type for GeneralCategory

GraphemeClusterBreak

Enumerated property Grapheme_Cluster_Break.

GraphemeClusterBreakULE

ULE type for GraphemeClusterBreak

HangulSyllableType

Enumerated property Hangul_Syllable_Type

HangulSyllableTypeULE

ULE type for HangulSyllableType

IndicSyllabicCategory

Property Indic_Syllabic_Category. See UAX #44: https://www.unicode.org/reports/tr44/#Indic_Syllabic_Category.

IndicSyllabicCategoryULE

ULE type for IndicSyllabicCategory

JoiningType

Enumerated property Joining_Type. See Section 9.2, Arabic Cursive Joining in The Unicode Standard for the summary of each property value.

JoiningTypeULE

ULE type for JoiningType

LineBreak

Enumerated property Line_Break.

LineBreakULE

ULE type for LineBreak

PropertyEnumToValueNameLinearMapper

A struct capable of looking up a property name from a value Access its data by calling Self::as_borrowed() and using the methods on PropertyEnumToValueNameLinearMapperBorrowed.

PropertyEnumToValueNameLinearMapperBorrowed

A borrowed wrapper around property value name-to-enum data, returned by PropertyEnumToValueNameLinearMapper::as_borrowed(). More efficient to query.

PropertyEnumToValueNameLinearTiny4Mapper

A struct capable of looking up a property name from a value Access its data by calling Self::as_borrowed() and using the methods on PropertyEnumToValueNameLinearTiny4MapperBorrowed.

PropertyEnumToValueNameLinearTiny4MapperBorrowed

A borrowed wrapper around property value name-to-enum data, returned by PropertyEnumToValueNameLinearTiny4Mapper::as_borrowed(). More efficient to query.

PropertyEnumToValueNameSparseMapper

A struct capable of looking up a property name from a value Access its data by calling Self::as_borrowed() and using the methods on PropertyEnumToValueNameSparseMapperBorrowed.

PropertyEnumToValueNameSparseMapperBorrowed

A borrowed wrapper around property value name-to-enum data, returned by PropertyEnumToValueNameSparseMapper::as_borrowed(). More efficient to query.

PropertyValueNameToEnumMapper

A struct capable of looking up a property value from a string name. Access its data by calling Self::as_borrowed() and using the methods on PropertyValueNameToEnumMapperBorrowed.

PropertyValueNameToEnumMapperBorrowed

A borrowed wrapper around property value name-to-enum data, returned by PropertyValueNameToEnumMapper::as_borrowed(). More efficient to query.

Script

Enumerated property Script.

ScriptULE

ULE type for Script

SentenceBreak

Enumerated property Sentence_Break. See “Default Sentence Boundary Specification” in UAX #29 for the summary of each property value: https://www.unicode.org/reports/tr29/#Default_Word_Boundaries.

SentenceBreakULE

ULE type for SentenceBreak

WordBreak

Enumerated property Word_Break.

WordBreakULE

ULE type for WordBreak

Enums

GeneralCategory

Enumerated property General_Category.

Functions

get_loose_u16 🔒

Avoid monomorphizing multiple copies of this function

get_strict_u16 🔒

Avoid monomorphizing multiple copies of this function