icu_properties::props

Trait BinaryProperty

pub trait BinaryProperty: Sealed + Sized {
    const NAME: &'static [u8];
    const SHORT_NAME: &'static [u8];

    // Provided method
    fn for_char(ch: char) -> bool { ... }
}

A binary Unicode character property.

The descriptions of most properties are taken from TR44, the documentation for the Unicode Character Database. Some properties are instead defined in TR18, the documentation for Unicode regular expressions. In particular, Annex C of this document defines properties for POSIX compatibility.

๐ซ This trait is sealed; it cannot be implemented by user code. If an API requests an item that implements this trait, please consider using a type from the implementors listed below.

Required Associated Constants

const NAME: &'static [u8]

The name of this property

const SHORT_NAME: &'static [u8]

The abbreviated name of this property, if it exists, otherwise the name

Provided Methods

fn for_char(ch: char) -> bool

Convenience method for CodePointSetData::new().contains(ch)

โจ Enabled with the compiled_data Cargo feature.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.