icu_properties

Module provider

🚧 [Unstable] Data provider struct definitions for this ICU4X component.

🚧 This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Read more about data providers: icu_provider

Re-exports

pub use names::PropertyNameLongBidiClassV1;

pub use names::PropertyNameLongCanonicalCombiningClassV1;

pub use names::PropertyNameLongEastAsianWidthV1;

pub use names::PropertyNameLongGeneralCategoryV1;

pub use names::PropertyNameLongGraphemeClusterBreakV1;

pub use names::PropertyNameLongHangulSyllableTypeV1;

pub use names::PropertyNameLongIndicSyllabicCategoryV1;

pub use names::PropertyNameLongJoiningTypeV1;

pub use names::PropertyNameLongLineBreakV1;

pub use names::PropertyNameLongScriptV1;

pub use names::PropertyNameLongSentenceBreakV1;

pub use names::PropertyNameLongVerticalOrientationV1;

pub use names::PropertyNameLongWordBreakV1;

pub use names::PropertyNameParseBidiClassV1;

pub use names::PropertyNameParseCanonicalCombiningClassV1;

pub use names::PropertyNameParseEastAsianWidthV1;

pub use names::PropertyNameParseGeneralCategoryMaskV1;

pub use names::PropertyNameParseGeneralCategoryV1;

pub use names::PropertyNameParseGraphemeClusterBreakV1;

pub use names::PropertyNameParseHangulSyllableTypeV1;

pub use names::PropertyNameParseIndicSyllabicCategoryV1;

pub use names::PropertyNameParseJoiningTypeV1;

pub use names::PropertyNameParseLineBreakV1;

pub use names::PropertyNameParseScriptV1;

pub use names::PropertyNameParseSentenceBreakV1;

pub use names::PropertyNameParseVerticalOrientationV1;

pub use names::PropertyNameParseWordBreakV1;

pub use names::PropertyNameShortBidiClassV1;

pub use names::PropertyNameShortCanonicalCombiningClassV1;

pub use names::PropertyNameShortEastAsianWidthV1;

pub use names::PropertyNameShortGeneralCategoryV1;

pub use names::PropertyNameShortGraphemeClusterBreakV1;

pub use names::PropertyNameShortHangulSyllableTypeV1;

pub use names::PropertyNameShortIndicSyllabicCategoryV1;

pub use names::PropertyNameShortJoiningTypeV1;

pub use names::PropertyNameShortLineBreakV1;

pub use names::PropertyNameShortScriptV1;

pub use names::PropertyNameShortSentenceBreakV1;

pub use names::PropertyNameShortVerticalOrientationV1;

pub use names::PropertyNameShortWordBreakV1;

Modules

names

🚧 [Unstable] Property names-related data for this component

Structs

Baked

Baked data

GeneralCategoryULE

ULE type for GeneralCategory

PropertyBinaryAlnumV1

PropertyBinaryAlnumV1

PropertyBinaryAlphabeticV1

PropertyBinaryAlphabeticV1

PropertyBinaryAsciiHexDigitV1

PropertyBinaryAsciiHexDigitV1

PropertyBinaryBasicEmojiV1

PropertyBinaryBasicEmojiV1

PropertyBinaryBidiControlV1

PropertyBinaryBidiControlV1

PropertyBinaryBidiMirroredV1

PropertyBinaryBidiMirroredV1

PropertyBinaryBlankV1

PropertyBinaryBlankV1

PropertyBinaryCaseIgnorableV1

PropertyBinaryCaseIgnorableV1

PropertyBinaryCaseSensitiveV1

PropertyBinaryCaseSensitiveV1

PropertyBinaryCasedV1

PropertyBinaryCasedV1

PropertyBinaryChangesWhenCasefoldedV1

PropertyBinaryChangesWhenCasefoldedV1

PropertyBinaryChangesWhenCasemappedV1

PropertyBinaryChangesWhenCasemappedV1

PropertyBinaryChangesWhenLowercasedV1

PropertyBinaryChangesWhenLowercasedV1

PropertyBinaryChangesWhenNfkcCasefoldedV1

PropertyBinaryChangesWhenNfkcCasefoldedV1

PropertyBinaryChangesWhenTitlecasedV1

PropertyBinaryChangesWhenTitlecasedV1

PropertyBinaryChangesWhenUppercasedV1

PropertyBinaryChangesWhenUppercasedV1

PropertyBinaryDashV1

PropertyBinaryDashV1

PropertyBinaryDefaultIgnorableCodePointV1

PropertyBinaryDefaultIgnorableCodePointV1

PropertyBinaryDeprecatedV1

PropertyBinaryDeprecatedV1

PropertyBinaryDiacriticV1

PropertyBinaryDiacriticV1

PropertyBinaryEmojiComponentV1

PropertyBinaryEmojiComponentV1

PropertyBinaryEmojiModifierBaseV1

PropertyBinaryEmojiModifierBaseV1

PropertyBinaryEmojiModifierV1

PropertyBinaryEmojiModifierV1

PropertyBinaryEmojiPresentationV1

PropertyBinaryEmojiPresentationV1

PropertyBinaryEmojiV1

PropertyBinaryEmojiV1

PropertyBinaryExtendedPictographicV1

PropertyBinaryExtendedPictographicV1

PropertyBinaryExtenderV1

PropertyBinaryExtenderV1

PropertyBinaryFullCompositionExclusionV1

PropertyBinaryFullCompositionExclusionV1

PropertyBinaryGraphV1

PropertyBinaryGraphV1

PropertyBinaryGraphemeBaseV1

PropertyBinaryGraphemeBaseV1

PropertyBinaryGraphemeExtendV1

PropertyBinaryGraphemeExtendV1

PropertyBinaryGraphemeLinkV1

PropertyBinaryGraphemeLinkV1

PropertyBinaryHexDigitV1

PropertyBinaryHexDigitV1

PropertyBinaryHyphenV1

PropertyBinaryHyphenV1

PropertyBinaryIdContinueV1

PropertyBinaryIdContinueV1

PropertyBinaryIdStartV1

PropertyBinaryIdStartV1

PropertyBinaryIdeographicV1

PropertyBinaryIdeographicV1

PropertyBinaryIdsBinaryOperatorV1

PropertyBinaryIdsBinaryOperatorV1

PropertyBinaryIdsTrinaryOperatorV1

PropertyBinaryIdsTrinaryOperatorV1

PropertyBinaryJoinControlV1

PropertyBinaryJoinControlV1

PropertyBinaryLogicalOrderExceptionV1

PropertyBinaryLogicalOrderExceptionV1

PropertyBinaryLowercaseV1

PropertyBinaryLowercaseV1

PropertyBinaryMathV1

PropertyBinaryMathV1

PropertyBinaryNfcInertV1

PropertyBinaryNfcInertV1

PropertyBinaryNfdInertV1

PropertyBinaryNfdInertV1

PropertyBinaryNfkcInertV1

PropertyBinaryNfkcInertV1

PropertyBinaryNfkdInertV1

PropertyBinaryNfkdInertV1

PropertyBinaryNoncharacterCodePointV1

PropertyBinaryNoncharacterCodePointV1

PropertyBinaryPatternSyntaxV1

PropertyBinaryPatternSyntaxV1

PropertyBinaryPatternWhiteSpaceV1

PropertyBinaryPatternWhiteSpaceV1

PropertyBinaryPrependedConcatenationMarkV1

PropertyBinaryPrependedConcatenationMarkV1

PropertyBinaryPrintV1

PropertyBinaryPrintV1

PropertyBinaryQuotationMarkV1

PropertyBinaryQuotationMarkV1

PropertyBinaryRadicalV1

PropertyBinaryRadicalV1

PropertyBinaryRegionalIndicatorV1

PropertyBinaryRegionalIndicatorV1

PropertyBinarySegmentStarterV1

PropertyBinarySegmentStarterV1

PropertyBinarySentenceTerminalV1

PropertyBinarySentenceTerminalV1

PropertyBinarySoftDottedV1

PropertyBinarySoftDottedV1

PropertyBinaryTerminalPunctuationV1

PropertyBinaryTerminalPunctuationV1

PropertyBinaryUnifiedIdeographV1

PropertyBinaryUnifiedIdeographV1

PropertyBinaryUppercaseV1

PropertyBinaryUppercaseV1

PropertyBinaryVariationSelectorV1

PropertyBinaryVariationSelectorV1

PropertyBinaryWhiteSpaceV1

PropertyBinaryWhiteSpaceV1

PropertyBinaryXdigitV1

PropertyBinaryXdigitV1

PropertyBinaryXidContinueV1

PropertyBinaryXidContinueV1

PropertyBinaryXidStartV1

PropertyBinaryXidStartV1

PropertyEnumBidiClassV1

Data marker for the ‘BidiClass’ Unicode property

PropertyEnumBidiMirroringGlyphV1

Data marker for the ‘BidiMirroringGlyph’ Unicode property

PropertyEnumCanonicalCombiningClassV1

Data marker for the ‘CanonicalCombiningClass’ Unicode property

PropertyEnumEastAsianWidthV1

Data marker for the ‘EastAsianWidth’ Unicode property

PropertyEnumGeneralCategoryV1

Data marker for the ‘GeneralCategory’ Unicode property

PropertyEnumGraphemeClusterBreakV1

Data marker for the ‘GraphemeClusterBreak’ Unicode property

PropertyEnumHangulSyllableTypeV1

Data marker for the ‘HangulSyllableType’ Unicode property

PropertyEnumIndicConjunctBreakV1

Data marker for the ‘IndicConjunctBreak’ Unicode property

PropertyEnumIndicSyllabicCategoryV1

Data marker for the ‘IndicSyllabicCategory’ Unicode property

PropertyEnumJoiningTypeV1

Data marker for the ‘JoiningType’ Unicode property

PropertyEnumLineBreakV1

Data marker for the ‘LineBreak’ Unicode property

PropertyEnumScriptV1

Data marker for the ‘Script’ Unicode property

PropertyEnumSentenceBreakV1

Data marker for the ‘SentenceBreak’ Unicode property

PropertyEnumVerticalOrientationV1

Data marker for the ‘Vertical_Orientation’ Unicode property

PropertyEnumWordBreakV1

Data marker for the ‘WordBreak’ Unicode property

PropertyScriptWithExtensionsV1

PropertyScriptWithExtensionsV1

ScriptWithExtensionsProperty

A struct that efficiently stores Script and Script_Extensions property data.

Enums

PropertyCodePointMap

A map efficiently storing data about individual characters.

PropertyCodePointSet

A set of characters which share a particular property value.

PropertyUnicodeSet

A set of characters and strings which share a particular property value.

Constants

MARKERS

All data keys in this module.