Module icu_properties::provider::key

Resource keys for icu_uniset

Constants

ALL_MAP_KEYS

The set of all resource keys supported by icu_uniset.

ALL_SCRIPT_EXTENSIONS_KEYS

The set of all resource keys supported by icu_uniset.

ALL_SET_KEYS

The set of all resource keys supported by icu_uniset.

ALNUM_V1

ALPHABETIC_V1

ASCII_HEX_DIGIT_V1

BIDI_CONTROL_V1

BIDI_MIRRORED_V1

BLANK_V1

CANONICAL_COMBINING_CLASS_V1

CASED_V1

CASE_IGNORABLE_V1

CASE_SENSITIVE_V1

CHANGES_WHEN_CASEFOLDED_V1

CHANGES_WHEN_CASEMAPPED_V1

CHANGES_WHEN_LOWERCASED_V1

CHANGES_WHEN_NFKC_CASEFOLDED_V1

CHANGES_WHEN_TITLECASED_V1

CHANGES_WHEN_UPPERCASED_V1

DASH_V1

DEFAULT_IGNORABLE_CODE_POINT_V1

DEPRECATED_V1

DIACRITIC_V1

EAST_ASIAN_WIDTH_V1

EMOJI_COMPONENT_V1

EMOJI_MODIFIER_BASE_V1

EMOJI_MODIFIER_V1

EMOJI_PRESENTATION_V1

EMOJI_V1

EXTENDED_PICTOGRAPHIC_V1

EXTENDER_V1

FULL_COMPOSITION_EXCLUSION_V1

GENERAL_CATEGORY_CASED_LETTER_V1

GENERAL_CATEGORY_CLOSE_PUNCTUATION_V1

GENERAL_CATEGORY_CONNECTOR_PUNCTUATION_V1

GENERAL_CATEGORY_CONTROL_V1

GENERAL_CATEGORY_CURRENCY_SYMBOL_V1

GENERAL_CATEGORY_DASH_PUNCTUATION_V1

GENERAL_CATEGORY_DIGIT_V1

GENERAL_CATEGORY_ENCLOSING_MARK_V1

GENERAL_CATEGORY_FINAL_PUNCTUATION_V1

GENERAL_CATEGORY_FORMAT_V1

GENERAL_CATEGORY_INITIAL_PUNCTUATION_V1

GENERAL_CATEGORY_LETTER_NUMBER_V1

GENERAL_CATEGORY_LETTER_V1

GENERAL_CATEGORY_LINE_SEPARATOR_V1

GENERAL_CATEGORY_LOWERCASE_LETTER_V1

GENERAL_CATEGORY_MARK_V1

GENERAL_CATEGORY_MATH_SYMBOL_V1

GENERAL_CATEGORY_MODIFIER_LETTER_V1

GENERAL_CATEGORY_MODIFIER_SYMBOL_V1

GENERAL_CATEGORY_NONSPACING_MARK_V1

GENERAL_CATEGORY_NUMBER_V1

GENERAL_CATEGORY_OPEN_PUNCTUATION_V1

GENERAL_CATEGORY_OTHER_LETTER_V1

GENERAL_CATEGORY_OTHER_NUMBER_V1

GENERAL_CATEGORY_OTHER_PUNCTUATION_V1

GENERAL_CATEGORY_OTHER_SYMBOL_V1

GENERAL_CATEGORY_OTHER_V1

GENERAL_CATEGORY_PARAGRAPH_SEPARATOR_V1

GENERAL_CATEGORY_PRIVATE_USE_V1

GENERAL_CATEGORY_PUNCTUATION_V1

GENERAL_CATEGORY_SEPARATOR_V1

GENERAL_CATEGORY_SPACE_SEPARATOR_V1

GENERAL_CATEGORY_SPACING_MARK_V1

GENERAL_CATEGORY_SURROGATE_V1

GENERAL_CATEGORY_SYMBOL_V1

GENERAL_CATEGORY_TITLECASE_LETTER_V1

GENERAL_CATEGORY_UNASSIGNED_V1

GENERAL_CATEGORY_UPPERCASE_LETTER_V1

GENERAL_CATEGORY_V1

GRAPHEME_BASE_V1

GRAPHEME_CLUSTER_BREAK_V1

GRAPHEME_EXTEND_V1

GRAPHEME_LINK_V1

GRAPH_V1

HEX_DIGIT_V1

HYPHEN_V1

IDEOGRAPHIC_V1

IDS_BINARY_OPERATOR_V1

IDS_TRINARY_OPERATOR_V1

ID_CONTINUE_V1

ID_START_V1

JOIN_CONTROL_V1

LINE_BREAK_V1

LOGICAL_ORDER_EXCEPTION_V1

LOWERCASE_V1

MATH_V1

NFC_INERT_V1

NFD_INERT_V1

NFKC_INERT_V1

NFKD_INERT_V1

NONCHARACTER_CODE_POINT_V1

PATTERN_SYNTAX_V1

PATTERN_WHITE_SPACE_V1

PREPENDED_CONCATENATION_MARK_V1

PRINT_V1

QUOTATION_MARK_V1

RADICAL_V1

REGIONAL_INDICATOR_V1

SCRIPT_ADLAM_V1

SCRIPT_AHOM_V1

SCRIPT_ANATOLIAN_HIEROGLYPHS_V1

SCRIPT_ARABIC_V1

SCRIPT_ARMENIAN_V1

SCRIPT_AVESTAN_V1

SCRIPT_BALINESE_V1

SCRIPT_BAMUM_V1

SCRIPT_BASSA_VAH_V1

SCRIPT_BATAK_V1

SCRIPT_BENGALI_V1

SCRIPT_BHAIKSUKI_V1

SCRIPT_BOPOMOFO_V1

SCRIPT_BRAHMI_V1

SCRIPT_BRAILLE_V1

SCRIPT_BUGINESE_V1

SCRIPT_BUHID_V1

SCRIPT_CANADIAN_ABORIGINAL_V1

SCRIPT_CARIAN_V1

SCRIPT_CAUCASIAN_ALBANIAN_V1

SCRIPT_CHAKMA_V1

SCRIPT_CHAM_V1

SCRIPT_CHEROKEE_V1

SCRIPT_CHORASMIAN_V1

SCRIPT_COMMON_V1

SCRIPT_COPTIC_V1

SCRIPT_CUNEIFORM_V1

SCRIPT_CYPRIOT_V1

SCRIPT_CYPRO_MINOAN_V1

SCRIPT_CYRILLIC_V1

SCRIPT_DESERET_V1

SCRIPT_DEVANAGARI_V1

SCRIPT_DIVES_AKURU_V1

SCRIPT_DOGRA_V1

SCRIPT_DUPLOYAN_V1

SCRIPT_EGYPTIAN_HIEROGLYPHS_V1

SCRIPT_ELBASAN_V1

SCRIPT_ELYMAIC_V1

SCRIPT_ETHIOPIC_V1

SCRIPT_EXTENSIONS_V1

SCRIPT_GEORGIAN_V1

SCRIPT_GLAGOLITIC_V1

SCRIPT_GOTHIC_V1

SCRIPT_GRANTHA_V1

SCRIPT_GREEK_V1

SCRIPT_GUJARATI_V1

SCRIPT_GUNJALA_GONDI_V1

SCRIPT_GURMUKHI_V1

SCRIPT_HANGUL_V1

SCRIPT_HANIFI_ROHINGYA_V1

SCRIPT_HANUNOO_V1

SCRIPT_HAN_V1

SCRIPT_HATRAN_V1

SCRIPT_HEBREW_V1

SCRIPT_HIRAGANA_V1

SCRIPT_IMPERIAL_ARAMAIC_V1

SCRIPT_INHERITED_V1

SCRIPT_INSCRIPTIONAL_PAHLAVI_V1

SCRIPT_INSCRIPTIONAL_PARTHIAN_V1

SCRIPT_JAVANESE_V1

SCRIPT_KAITHI_V1

SCRIPT_KANNADA_V1

SCRIPT_KATAKANA_V1

SCRIPT_KAYAH_LI_V1

SCRIPT_KHAROSHTHI_V1

SCRIPT_KHITAN_SMALL_SCRIPT_V1

SCRIPT_KHMER_V1

SCRIPT_KHOJKI_V1

SCRIPT_KHUDAWADI_V1

SCRIPT_LAO_V1

SCRIPT_LATIN_V1

SCRIPT_LEPCHA_V1

SCRIPT_LIMBU_V1

SCRIPT_LINEAR_A_V1

SCRIPT_LINEAR_B_V1

SCRIPT_LISU_V1

SCRIPT_LYCIAN_V1

SCRIPT_LYDIAN_V1

SCRIPT_MAHAJANI_V1

SCRIPT_MAKASAR_V1

SCRIPT_MALAYALAM_V1

SCRIPT_MANDAIC_V1

SCRIPT_MANICHAEAN_V1

SCRIPT_MARCHEN_V1

SCRIPT_MASARAM_GONDI_V1

SCRIPT_MEDEFAIDRIN_V1

SCRIPT_MEETEI_MAYEK_V1

SCRIPT_MENDE_KIKAKUI_V1

SCRIPT_MEROITIC_CURSIVE_V1

SCRIPT_MEROITIC_HIEROGLYPHS_V1

SCRIPT_MIAO_V1

SCRIPT_MODI_V1

SCRIPT_MONGOLIAN_V1

SCRIPT_MRO_V1

SCRIPT_MULTANI_V1

SCRIPT_MYANMAR_V1

SCRIPT_NABATAEAN_V1

SCRIPT_NANDINAGARI_V1

SCRIPT_NEWA_V1

SCRIPT_NEW_TAI_LUE_V1

SCRIPT_NKO_V1

SCRIPT_NUSHU_V1

SCRIPT_NYIAKENG_PUACHUE_HMONG_V1

SCRIPT_OGHAM_V1

SCRIPT_OLD_HUNGARIAN_V1

SCRIPT_OLD_ITALIC_V1

SCRIPT_OLD_NORTH_ARABIAN_V1

SCRIPT_OLD_PERMIC_V1

SCRIPT_OLD_PERSIAN_V1

SCRIPT_OLD_SOGDIAN_V1

SCRIPT_OLD_SOUTH_ARABIAN_V1

SCRIPT_OLD_TURKIC_V1

SCRIPT_OLD_UYGHUR_V1

SCRIPT_OL_CHIKI_V1

SCRIPT_ORIYA_V1

SCRIPT_OSAGE_V1

SCRIPT_OSMANYA_V1

SCRIPT_PAHAWH_HMONG_V1

SCRIPT_PALMYRENE_V1

SCRIPT_PAU_CIN_HAU_V1

SCRIPT_PHAGS_PA_V1

SCRIPT_PHOENICIAN_V1

SCRIPT_PSALTER_PAHLAVI_V1

SCRIPT_REJANG_V1

SCRIPT_RUNIC_V1

SCRIPT_SAMARITAN_V1

SCRIPT_SAURASHTRA_V1

SCRIPT_SHARADA_V1

SCRIPT_SHAVIAN_V1

SCRIPT_SIDDHAM_V1

SCRIPT_SIGNWRITING_V1

SCRIPT_SINHALA_V1

SCRIPT_SOGDIAN_V1

SCRIPT_SORA_SOMPENG_V1

SCRIPT_SOYOMBO_V1

SCRIPT_SUNDANESE_V1

SCRIPT_SYLOTI_NAGRI_V1

SCRIPT_SYRIAC_V1

SCRIPT_TAGALOG_V1

SCRIPT_TAGBANWA_V1

SCRIPT_TAI_LE_V1

SCRIPT_TAI_THAM_V1

SCRIPT_TAI_VIET_V1

SCRIPT_TAKRI_V1

SCRIPT_TAMIL_V1

SCRIPT_TANGSA_V1

SCRIPT_TANGUT_V1

SCRIPT_TELUGU_V1

SCRIPT_THAANA_V1

SCRIPT_THAI_V1

SCRIPT_TIBETAN_V1

SCRIPT_TIFINAGH_V1

SCRIPT_TIRHUTA_V1

SCRIPT_TOTO_V1

SCRIPT_UGARITIC_V1

SCRIPT_UNKNOWN_V1

SCRIPT_V1

SCRIPT_VAI_V1

SCRIPT_VITHKUQI_V1

SCRIPT_WANCHO_V1

SCRIPT_WARANG_CITI_V1

SCRIPT_YEZIDI_V1

SCRIPT_YI_V1

SCRIPT_ZANABAZAR_SQUARE_V1

SEGMENT_STARTER_V1

SENTENCE_BREAK_V1

SENTENCE_TERMINAL_V1

SOFT_DOTTED_V1

TERMINAL_PUNCTUATION_V1

UNIFIED_IDEOGRAPH_V1

UPPERCASE_V1

VARIATION_SELECTOR_V1

WHITE_SPACE_V1

WORD_BREAK_V1

XDIGIT_V1

XID_CONTINUE_V1

XID_START_V1