Module icu_properties::provider::names

🚧 [Unstable] Property names-related data for this component

🚧 This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Read more about data providers: icu_provider

Structs

GeneralCategoryMaskNameToValueV1Marker

Marker type for PropertyValueNameToEnumMapV1: “propnames/from/gcm@1”

NormalizedPropertyNameStr

This is a property name that can be “loose matched” as according to PropertyValueAliases.txt

PropertyEnumToValueNameLinearMapV1

A mapping of property values to their names. A single instance of this map will only cover either long or short names, determined whilst loading data.

PropertyEnumToValueNameLinearTiny4MapV1

A mapping of property values to their names. A single instance of this map will only cover either long or short names, determined whilst loading data.

PropertyEnumToValueNameSparseMapV1

A mapping of property values to their names. A single instance of this map will only cover either long or short names, determined whilst loading data.

PropertyValueNameToEnumMapV1

A set of characters and strings which share a particular property value.