Struct icu_properties::runtime::UnicodeProperty

pub struct UnicodeProperty(pub u32);

This type can represent any Unicode property.

This is intended to be used in situations where the exact unicode property needed is only known at runtime, for example in regex engines.

The values are intended to be identical to ICU4C’s UProperty enum

Tuple Fields
0: u32
Implementations

impl UnicodeProperty

pub const Alphabetic: Self = _

Binary property Alphabetic

pub const AsciiHexDigit: Self = _

Binary property ASCII_Hex_Digit

pub const BidiControl: Self = _

Binary property Bidi_Control

pub const BidiMirrored: Self = _

Binary property Bidi_Mirrored

pub const Dash: Self = _

Binary property Dash

pub const DefaultIgnorableCodePoint: Self = _

Binary property Default_Ignorable_Code_Point

pub const Deprecated: Self = _

Binary property Deprecated

pub const Diacritic: Self = _

Binary property Diacritic

pub const Extender: Self = _

Binary property Extender

pub const FullCompositionExclusion: Self = _

Binary property Full_Composition_Exclusion

pub const GraphemeBase: Self = _

Binary property Grapheme_Base

pub const GraphemeExtend: Self = _

Binary property Grapheme_Extend

pub const GraphemeLink: Self = _

Binary property Grapheme_Link

pub const HexDigit: Self = _

Binary property Hex_Digit

pub const Hyphen: Self = _

Binary property Hyphen

pub const IdContinue: Self = _

Binary property ID_Continue

pub const IdStart: Self = _

Binary property ID_Start

pub const Ideographic: Self = _

Binary property Ideographic

pub const IdsBinaryOperator: Self = _

Binary property IDS_Binary_Operator

pub const IdsTrinaryOperator: Self = _

Binary property IDS_Trinary_Operator

pub const JoinControl: Self = _

Binary property Join_Control

pub const LogicalOrderException: Self = _

Binary property Logical_Order_Exception

pub const Lowercase: Self = _

Binary property Lowercase

pub const Math: Self = _

Binary property Math

pub const NoncharacterCodePoint: Self = _

Binary property Noncharacter_Code_Point

pub const QuotationMark: Self = _

Binary property Quotation_Mark

pub const Radical: Self = _

Binary property Radical

pub const SoftDotted: Self = _

Binary property Soft_Dotted

pub const TerminalPunctuation: Self = _

Binary property Terminal_Punctuation

pub const UnifiedIdeograph: Self = _

Binary property Unified_Ideograph

pub const Uppercase: Self = _

Binary property Uppercase

pub const WhiteSpace: Self = _

Binary property White_Space

pub const XidContinue: Self = _

Binary property XID_Continue

pub const XidStart: Self = _

Binary property XID_Start

pub const CaseSensitive: Self = _

Binary property Case_Sensitive

pub const SentenceTerminal: Self = _

Binary property Sentence_Terminal

pub const VariationSelector: Self = _

Binary property Variation_Selector

pub const NfdInert: Self = _

Binary property NFD_Inert

pub const NfkdInert: Self = _

Binary property NFKD_Inert

pub const NfcInert: Self = _

Binary property NFC_Inert

pub const NfkcInert: Self = _

Binary property NFKC_Inert

pub const SegmentStarter: Self = _

Binary property Segment_Starter

pub const PatternSyntax: Self = _

Binary property Pattern_Syntax

pub const PatternWhiteSpace: Self = _

Binary property Pattern_White_Space

pub const Alnum: Self = _

Binary property alnum

pub const Blank: Self = _

Binary property blank

pub const Graph: Self = _

Binary property graph

pub const Print: Self = _

Binary property print

pub const XDigit: Self = _

Binary property xdigit

pub const Cased: Self = _

Binary property Cased

pub const CaseIgnorable: Self = _

Binary property Case_Ignorable

pub const ChangesWhenLowercased: Self = _

Binary property Changes_When_Lowercased

pub const ChangesWhenUppercased: Self = _

Binary property Changes_When_Uppercased

pub const ChangesWhenTitlecased: Self = _

Binary property Changes_When_Titlecased

pub const ChangesWhenCasefolded: Self = _

Binary property Changes_When_Casefolded

pub const ChangesWhenCasemapped: Self = _

Binary property Changes_When_Casemapped

pub const ChangesWhenNfkcCasefolded: Self = _

Binary property Changes_When_NFKC_Casefolded

pub const Emoji: Self = _

Binary property Emoji

pub const EmojiPresentation: Self = _

Binary property Emoji_Presentation

pub const EmojiModifier: Self = _

Binary property Emoji_Modifier

pub const EmojiModifierBase: Self = _

Binary property Emoji_Modifier_Base

pub const EmojiComponent: Self = _

Binary property Emoji_Component

pub const RegionalIndicator: Self = _

Binary property Regional_Indicator

pub const PrependedConcatenationMark: Self = _

Binary property Prepended_Concatenation_Mark

pub const ExtendedPictographic: Self = _

Binary property Extended_Pictographic

pub const BasicEmoji: Self = _

Binary property Basic_Emoji

pub const EmojiKeycapSequence: Self = _

Binary property Emoji_Keycap_Sequence

pub const RgiEmojiModifierSequence: Self = _

Binary property RGI_Emoji_Modifier_Sequence

pub const RgiEmojiFlagSequence: Self = _

Binary property RGI_Emoji_Flag_Sequence

pub const RgiEmojiTagSequence: Self = _

Binary property RGI_Emoji_Tag_Sequence

pub const RgiEmojiZWJSequence: Self = _

Binary property RGI_Emoji_ZWJ_Sequence

pub const RgiEmoji: Self = _

Binary property RGI_Emoji

const BINARY_MAX: Self = Self::RgiEmoji

pub const BidiClass: Self = _

Enumerated property Bidi_Class

pub const Block: Self = _

Enumerated property Block

pub const CombiningClass: Self = _

Enumerated property Canonical_Combining_Class

pub const DecompositionType: Self = _

Enumerated property Decomposition_Type

pub const EastAsianWidth: Self = _

Enumerated property East_Asian_Width

pub const GeneralCategory: Self = _

Enumerated property General_Category

pub const JoiningGroup: Self = _

Enumerated property Joining_Group

pub const JoiningType: Self = _

Enumerated property Joining_Type

pub const LineBreak: Self = _

Enumerated property Line_Break

pub const NumericType: Self = _

Enumerated property Numeric_Type

pub const Script: Self = _

Enumerated property Script

pub const HangulSyllableType: Self = _

Enumerated property Hangul_Syllable_Type

pub const NFDQuickCheck: Self = _

Enumerated property NFD_Quick_Check

pub const NFKDQuickCheck: Self = _

Enumerated property NFKD_Quick_Check

pub const NFCQuickCheck: Self = _

Enumerated property NFC_Quick_Check

pub const NFKCQuickCheck: Self = _

Enumerated property NFKC_Quick_Check

pub const LeadCanonicalCombiningClass: Self = _

Enumerated property Lead_Canonical_Combining_Class

pub const TrailCanonicalCombiningClass: Self = _

Enumerated property Trail_Canonical_Combining_Class

pub const GraphemeClusterBreak: Self = _

Enumerated property Grapheme_Cluster_Break

pub const SentenceBreak: Self = _

Enumerated property Sentence_Break

pub const WordBreak: Self = _

Enumerated property Word_Break

pub const BidiPairedBracketType: Self = _

Enumerated property Bidi_Paired_Bracket_Type

pub const IndicPositionalCategory: Self = _

Enumerated property Indic_Positional_Category

pub const IndicSyllabicCategory: Self = _

Enumerated property Indic_Syllabic_Category

pub const VerticalOrientation: Self = _

Enumerated property Vertical_Orientation

const ENUMERATED_MAX: Self = Self::VerticalOrientation

pub const GeneralCategoryMask: Self = _

Mask property General_Category_Mask

pub const NumericValue: Self = _

Double property Numeric_Value

pub const Age: Self = _

String property Age

pub const BidiMirroringGlyph: Self = _

String property Bidi_Mirroring_Glyph

pub const CaseFolding: Self = _

String property Case_Folding

pub const ISOComment: Self = _

String property ISO_Comment

pub const LowercaseMapping: Self = _

String property Lowercase_Mapping

pub const Name: Self = _

String property Name

pub const SimpleCaseFolding: Self = _

String property Simple_Case_Folding

pub const SimpleLowercaseMapping: Self = _

String property Simple_Lowercase_Mapping

pub const SimpleTitlecaseMapping: Self = _

String property Simple_Titlecase_Mapping

pub const SimpleUppercaseMapping: Self = _

String property Simple_Uppercase_Mapping

pub const TitlecaseMapping: Self = _

String property Titlecase_Mapping

pub const Unicode1_Name: Self = _

String property Unicode_1_Name

pub const UppercaseMapping: Self = _

String property Uppercase_Mapping

pub const BidiPairedBracket: Self = _

String property Bidi_Paired_Bracket

const STRING_MAX: Self = Self::BidiPairedBracket

pub const ScriptExtensions: Self = _

Misc property Script_Extensions

impl UnicodeProperty

pub fn parse_ecma262_name(name: &str) -> Option<Self>

Given a property name (long, short, or alias), returns the corresponding UnicodeProperty value for it provided it belongs to the subset relevant for ECMA262 regexes

Returns none if the name does not match any of the names in this subset. Performs strict matching of names.

If using this to implement an ECMA262-compliant regex engine, please note these caveats:

This only returns binary and enumerated properties, as well as Self::ScriptExtensions. Lookup can be performed sufficiently with [Self::load_ecma262_binary_property_unstable()], maps::load_general_category(), maps::load_script() and script::load_script_with_extensions_unstable().

This does not handle the Any, Assigned, or ASCII pseudoproperties, since they are not defined as properties.

Any can be expressed as the range [\u{0}-\u{10FFFF}]

Assigned can be expressed as the inverse of the set gc=Cn (i.e., \P{gc=Cn}).

ASCII can be expressed as the range [\u{0}-\u{7F}]

ECMA262 regexes transparently allow General_Category_Mask values for GeneralCategory. This method does not return Self::GeneralCategoryMask, and instead relies on the caller to use mask-related lookup functions where necessary.

ECMA262 regexes allow treating General_Category (and gcm) values as binary properties, e.g. you can do things like \p{Lu} as shortform for \p{gc=Lu}. This method does not do so since these are property values, not properties, but you can use GeneralCategory::get_name_to_enum_mapper() or GeneralCategoryGroup::get_name_to_enum_mapper() to handle this.