icu_properties

Module script

Data and APIs for supporting both Script and Script_Extensions property values in an efficient structure.

Structs

ScriptExtensionsSet

A struct that wraps a Script array, such as in the return value for get_script_extensions_val().

ScriptWithExtensions

A wrapper around script extensions data. Can be obtained via load_script_with_extensions_unstable() and related getters.

ScriptWithExtensionsBorrowed

A borrowed wrapper around script extension data, returned by ScriptWithExtensions::as_borrowed(). More efficient to query.

Functions

load_script_with_extensions_unstable

A version of script_with_extensions that uses custom data provided by a DataProvider.

load_script_with_extensions_with_any_provider

A version of script_with_extensions that uses custom data provided by an AnyProvider.

script_with_extensions

Returns a ScriptWithExtensionsBorrowed struct that represents the data for the Script and Script_Extensions properties.