icu_properties::script

Struct ScriptWithExtensions

pub struct ScriptWithExtensions { /* private fields */ }

A wrapper around script extensions data. Can be obtained via load_script_with_extensions_unstable() and related getters.

Most useful methods are on ScriptWithExtensionsBorrowed obtained by calling ScriptWithExtensions::as_borrowed()

Implementations

impl ScriptWithExtensions

pub fn as_borrowed(&self) -> ScriptWithExtensionsBorrowed<'_>

Construct a borrowed version of this type that can be queried.

This avoids a potential small underlying cost per API call (ex: contains()) by consolidating it up front.

pub fn from_data( data: DataPayload<ScriptWithExtensionsPropertyV1Marker>, ) -> Self

Construct a new one from loaded data

Typically it is preferable to use getters like load_script_with_extensions_unstable() instead