icu_properties

Struct PropertyNamesLong

pub struct PropertyNamesLong<T: NamedEnumeratedProperty> { /* private fields */ }

A struct capable of looking up a property name from a value Access its data by calling Self::as_borrowed() and using the methods on PropertyNamesLongBorrowed.

Example

use icu::properties::props::CanonicalCombiningClass;
use icu::properties::PropertyNamesLong;

let names = PropertyNamesLong::<CanonicalCombiningClass>::new();
assert_eq!(
    names.get(CanonicalCombiningClass::KanaVoicing),
    Some("Kana_Voicing")
);
assert_eq!(
    names.get(CanonicalCombiningClass::AboveLeft),
    Some("Above_Left")
);

Implementations

impl<T: NamedEnumeratedProperty> PropertyNamesLong<T>

pub fn new() -> PropertyNamesLongBorrowed<'static, T>

Creates a new instance of PropertyNamesLongBorrowed<T>.

โจ Enabled with the compiled_data Cargo feature.

๐ Help choosing a constructor

pub fn try_new_unstable( provider: &(impl DataProvider<T::DataMarkerLong> + ?Sized), ) -> Result<Self, DataError>

A version of Self::new that uses custom data provided by a DataProvider.

๐ Help choosing a constructor

โ๏ธ The bounds on provider may change over time, including in SemVer minor releases.

pub fn as_borrowed(&self) -> PropertyNamesLongBorrowed<'_, T>

Construct a borrowed version of this type that can be queried.

This avoids a potential small underlying cost per API call (like get_static()) by consolidating it up front.