icu_properties

Struct Script

#[repr(transparent)]
pub struct Script(pub u16);

Enumerated property Script.

This is used with both the Script and Script_Extensions Unicode properties. Each character is assigned a single Script, but characters that are used in a particular subset of scripts will be in more than one Script_Extensions set. For example, DEVANAGARI DIGIT NINE has Script=Devanagari, but is also in the Script_Extensions set for Dogra, Kaithi, and Mahajani.

For more information, see UAX #24: http://www.unicode.org/reports/tr24/. See UScriptCode in ICU4C.

Tuple Fields
0: u16
Implementations

impl Script

pub const Adlam: Script

pub const Ahom: Script

pub const AnatolianHieroglyphs: Script

pub const Arabic: Script

pub const Armenian: Script

pub const Avestan: Script

pub const Balinese: Script

pub const Bamum: Script

pub const BassaVah: Script

pub const Batak: Script

pub const Bengali: Script

pub const Bhaiksuki: Script

pub const Bopomofo: Script

pub const Brahmi: Script

pub const Braille: Script

pub const Buginese: Script

pub const Buhid: Script

pub const CanadianAboriginal: Script

pub const Carian: Script

pub const CaucasianAlbanian: Script

pub const Chakma: Script

pub const Cham: Script

pub const Cherokee: Script

pub const Chorasmian: Script

pub const Common: Script

pub const Coptic: Script

pub const Cuneiform: Script

pub const Cypriot: Script

pub const CyproMinoan: Script

pub const Cyrillic: Script

pub const Deseret: Script

pub const Devanagari: Script

pub const DivesAkuru: Script

pub const Dogra: Script

pub const Duployan: Script

pub const EgyptianHieroglyphs: Script

pub const Elbasan: Script

pub const Elymaic: Script

pub const Ethiopian: Script

pub const Georgian: Script

pub const Glagolitic: Script

pub const Gothic: Script

pub const Grantha: Script

pub const Greek: Script

pub const Gujarati: Script

pub const GunjalaGondi: Script

pub const Gurmukhi: Script

pub const Han: Script

pub const Hangul: Script

pub const HanifiRohingya: Script

pub const Hanunoo: Script

pub const Hatran: Script

pub const Hebrew: Script

pub const Hiragana: Script

pub const ImperialAramaic: Script

pub const Inherited: Script

pub const InscriptionalPahlavi: Script

pub const InscriptionalParthian: Script

pub const Javanese: Script

pub const Kaithi: Script

pub const Kannada: Script

pub const Katakana: Script

pub const Kawi: Script

pub const KayahLi: Script

pub const Kharoshthi: Script

pub const KhitanSmallScript: Script

pub const Khmer: Script

pub const Khojki: Script

pub const Khudawadi: Script

pub const Lao: Script

pub const Latin: Script

pub const Lepcha: Script

pub const Limbu: Script

pub const LinearA: Script

pub const LinearB: Script

pub const Lisu: Script

pub const Lycian: Script

pub const Lydian: Script

pub const Mahajani: Script

pub const Makasar: Script

pub const Malayalam: Script

pub const Mandaic: Script

pub const Manichaean: Script

pub const Marchen: Script

pub const MasaramGondi: Script

pub const Medefaidrin: Script

pub const MeeteiMayek: Script

pub const MendeKikakui: Script

pub const MeroiticCursive: Script

pub const MeroiticHieroglyphs: Script

pub const Miao: Script

pub const Modi: Script

pub const Mongolian: Script

pub const Mro: Script

pub const Multani: Script

pub const Myanmar: Script

pub const Nabataean: Script

pub const NagMundari: Script

pub const Nandinagari: Script

pub const Nastaliq: Script

pub const NewTaiLue: Script

pub const Newa: Script

pub const Nko: Script

pub const Nushu: Script

pub const NyiakengPuachueHmong: Script

pub const Ogham: Script

pub const OlChiki: Script

pub const OldHungarian: Script

pub const OldItalic: Script

pub const OldNorthArabian: Script

pub const OldPermic: Script

pub const OldPersian: Script

pub const OldSogdian: Script

pub const OldSouthArabian: Script

pub const OldTurkic: Script

pub const OldUyghur: Script

pub const Oriya: Script

pub const Osage: Script

pub const Osmanya: Script

pub const PahawhHmong: Script

pub const Palmyrene: Script

pub const PauCinHau: Script

pub const PhagsPa: Script

pub const Phoenician: Script

pub const PsalterPahlavi: Script

pub const Rejang: Script

pub const Runic: Script

pub const Samaritan: Script

pub const Saurashtra: Script

pub const Sharada: Script

pub const Shavian: Script

pub const Siddham: Script

pub const SignWriting: Script

pub const Sinhala: Script

pub const Sogdian: Script

pub const SoraSompeng: Script

pub const Soyombo: Script

pub const Sundanese: Script

pub const SylotiNagri: Script

pub const Syriac: Script

pub const Tagalog: Script

pub const Tagbanwa: Script

pub const TaiLe: Script

pub const TaiTham: Script

pub const TaiViet: Script

pub const Takri: Script

pub const Tamil: Script

pub const Tangsa: Script

pub const Tangut: Script

pub const Telugu: Script

pub const Thaana: Script

pub const Thai: Script

pub const Tibetan: Script

pub const Tifinagh: Script

pub const Tirhuta: Script

pub const Toto: Script

pub const Ugaritic: Script

pub const Unknown: Script

pub const Vai: Script

pub const Vithkuqi: Script

pub const Wancho: Script

pub const WarangCiti: Script

pub const Yezidi: Script

pub const Yi: Script

pub const ZanabazarSquare: Script

impl Script

pub const fn name_to_enum_mapper() -> PropertyValueNameToEnumMapperBorrowed<'static, Script>

Return a PropertyValueNameToEnumMapper, capable of looking up values from strings for the Script enumerated property.

โจ Enabled with the compiled_data Cargo feature.

๐ Help choosing a constructor

Example

use icu::properties::Script;

let lookup = Script::name_to_enum_mapper();
// short name for value
assert_eq!(lookup.get_strict("Brah"), Some(Script::Brahmi));
assert_eq!(lookup.get_strict("Hang"), Some(Script::Hangul));
// long name for value
assert_eq!(lookup.get_strict("Brahmi"), Some(Script::Brahmi));
assert_eq!(lookup.get_strict("Hangul"), Some(Script::Hangul));
// name has incorrect casing
assert_eq!(lookup.get_strict("brahmi"), None);
// loose matching of name
assert_eq!(lookup.get_loose("brahmi"), Some(Script::Brahmi));
// fake property
assert_eq!(lookup.get_strict("Linear_Z"), None);

pub fn get_name_to_enum_mapper( provider: &(impl DataProvider<ScriptNameToValueV1Marker> + ?Sized), ) -> Result<PropertyValueNameToEnumMapper<Script>, PropertiesError>

A version of Script::name_to_enum_mapper() that uses custom data provided by a DataProvider.

๐ Help choosing a constructor

pub const fn enum_to_short_name_mapper() -> PropertyEnumToValueNameLinearTiny4MapperBorrowed<'static, Script>

Return a PropertyEnumToValueNameLinearMapper, capable of looking up short names for values of the Script enumerated property.

โจ Enabled with the compiled_data Cargo feature.

๐ Help choosing a constructor

Example

use icu::properties::Script;
use tinystr::tinystr;

let lookup = Script::enum_to_short_name_mapper();
assert_eq!(lookup.get(Script::Brahmi), Some(tinystr!(4, "Brah")));
assert_eq!(lookup.get(Script::Hangul), Some(tinystr!(4, "Hang")));

pub fn get_enum_to_short_name_mapper( provider: &(impl DataProvider<ScriptValueToShortNameV1Marker> + ?Sized), ) -> Result<PropertyEnumToValueNameLinearTiny4Mapper<Script>, PropertiesError>

A version of Script::enum_to_short_name_mapper() that uses custom data provided by a DataProvider.

๐ Help choosing a constructor

pub const fn enum_to_long_name_mapper() -> PropertyEnumToValueNameLinearMapperBorrowed<'static, Script>

Return a PropertyEnumToValueNameLinearTiny4Mapper, capable of looking up long names for values of the Script enumerated property.

โจ Enabled with the compiled_data Cargo feature.

๐ Help choosing a constructor

Example

use icu::properties::Script;

let lookup = Script::enum_to_long_name_mapper();
assert_eq!(lookup.get(Script::Brahmi), Some("Brahmi"));
assert_eq!(lookup.get(Script::Hangul), Some("Hangul"));

pub fn get_enum_to_long_name_mapper( provider: &(impl DataProvider<ScriptValueToLongNameV1Marker> + ?Sized), ) -> Result<PropertyEnumToValueNameLinearMapper<Script>, PropertiesError>

A version of Script::enum_to_long_name_mapper() that uses custom data provided by a DataProvider.

๐ Help choosing a constructor